split javascript runtime start and persistence tests
