+++ pkg/services/factory_sessions/internal/execution/fake_service_runtime_test.go
@@ -3,7 +3,6 @@
 package factorysessionexecution
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
@@ -17,376 +16,12 @@ import (
 
 	"github.com/portpowered/infinite-you/internal/testutil/checkpointfixtures"
 	"github.com/portpowered/infinite-you/internal/testutil/factoryruntimefixtures"
-	platformfilesystem "github.com/portpowered/infinite-you/pkg/platform/filesystem"
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	"github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	"github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/execution/runtimepersist"
-	"github.com/portpowered/infinite-you/pkg/services/recordings"
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 )
-
-type javaScriptRuntimeServiceConfig struct {
-	ProjectRoot        string
-	ChildExecutorMode  string
-	InvocationExecutor workerexecution.InvocationExecutor
-	Persistence        runtimepersist.Store
-	Clock              factory.Clock
-	Workflows          factory.JavaScriptWorkflows
-}
-
-func testRuntimePersistenceStoreFactory(projectRoot string) (runtimepersist.Store, error) {
-	return runtimepersist.NewProjectStore(projectRoot, platformfilesystem.Local{})
-}
-
-func mustTestRuntimePersistenceStore(t *testing.T, dir string) runtimepersist.Store {
-	t.Helper()
-	store, err := runtimepersist.NewDirectoryStore(dir, platformfilesystem.Local{})
-	if err != nil {
-		t.Fatalf("NewDirectoryStore: %v", err)
-	}
-	return store
-}
-
-func newConfiguredJavaScriptRuntimeService(config javaScriptRuntimeServiceConfig) *JavaScriptRuntimeService {
-	workflows := config.Workflows
-	if workflows == nil {
-		workflows = factoryruntimefixtures.ScriptedJavaScriptWorkflows{}
-	}
-	clock := config.Clock
-	if clock == nil {
-		clock = durableFixedClock{now: time.Date(2026, 1, 1, 0, 0, 0, 0, time.UTC)}
-	}
-	return NewJavaScriptRuntimeService(
-		config.ProjectRoot, config.ChildExecutorMode, config.InvocationExecutor,
-		config.Persistence, clock, testSyncWaitScheduler{}, checkpointfixtures.CheckpointSummariesFixture{
-			BuildResult:  checkpointfixtures.ResumableCheckpointSummaryResult(),
-			LatestResult: checkpointfixtures.ResumableCheckpointSummaryResult(),
-		},
-		workflows, orchestrationJavaScriptFromWorkflows(workflows), workflows,
-		nil, factory.JavaScriptWorkerSettings{}, mustTestRecordingWriter(),
-		testSessionIDGenerator,
-		nil, nil,
-	)
-}
-
-func mustTestRecordingWriter() recordings.PortableRecordingWriter {
-	return portableRecordingTestWriter{}
-}
-
-type portableRecordingTestWriter struct{}
-
-func (portableRecordingTestWriter) Write(path string, value recordings.PortableRecording) error {
-	if err := recordings.ValidatePortableRecording(value); err != nil {
-		return err
-	}
-	data, err := json.MarshalIndent(value, "", "  ")
-	if err != nil {
-		return err
-	}
-	if err := os.MkdirAll(filepath.Dir(path), 0o700); err != nil {
-		return err
-	}
-	return os.WriteFile(path, append(data, '\n'), 0o600)
-}
-
-func seedRuntimeSessionWithRunningDispatch(
-	service *JavaScriptRuntimeService,
-	sessionID, dispatchID, label string,
-) error {
-	if service == nil {
-		return NewValidationError("service", "service is required")
-	}
-	id, err := NormalizeSessionID(sessionID)
-	if err != nil {
-		return err
-	}
-	dispatchID = strings.TrimSpace(dispatchID)
-	if dispatchID == "" {
-		return NewValidationError("dispatchId", "dispatchId is required")
-	}
-
-	now := service.now()
-	session := SessionReadResult{
-		SessionID:        id,
-		Status:           LifecycleStatusRunning,
-		OrchestratorKind: interfaces.OrchestratorKindJavaScript,
-		Phase:            "execute",
-		Lifecycle:        &LifecycleTimestamps{StartedAt: &now},
-		Links:            InspectionLinksForSession(id, true),
-		Progress: &ProgressCounts{
-			TotalDispatches:    1,
-			InFlightDispatches: 1,
-		},
-	}
-	result := ResultReadResult{
-		SessionID:     id,
-		SessionStatus: LifecycleStatusRunning,
-		ResultStatus:  ResultStatusNotReady,
-		Availability: &ResultAvailabilityDetail{
-			Reason:    "RESULT_NOT_READY",
-			Message:   "Session is still running.",
-			Retryable: true,
-		},
-	}
-	dispatches := []DispatchSummary{{
-		ID: dispatchID, Status: DispatchStatusRunning, Phase: "execute", Label: label,
-	}}
-	state := &runtimeSessionState{
-		session:    session,
-		result:     result,
-		dispatches: dispatches,
-		dispatchStatusTransitions: map[string][]DispatchStatus{
-			dispatchID: {DispatchStatusQueued, DispatchStatusRunning},
-		},
-	}
-	state.events = rebuildRuntimeSessionCanonicalEvents(state)
-	service.mu.Lock()
-	defer service.mu.Unlock()
-	service.sessions[id] = state
-	return nil
-}
-
-func applyRuntimeTerminalOutcome(
-	service *JavaScriptRuntimeService,
-	sessionID string,
-	outcome factory.JavaScriptRuntimeOutcome,
-) error {
-	if service == nil {
-		return NewValidationError("service", "service is required")
-	}
-	id, err := NormalizeSessionID(sessionID)
-	if err != nil {
-		return err
-	}
-	service.mu.Lock()
-	defer service.mu.Unlock()
-	state, ok := service.sessions[id]
-	if !ok {
-		return ErrSessionNotFound
-	}
-	finishedAt := service.now()
-	terminal := runtimeSessionState{
-		session: cloneSessionRead(state.session),
-		result:  cloneResultRead(state.result),
-	}
-	if outcome.OK {
-		applyRuntimeSuccessProjection(&terminal, id, outcome, finishedAt)
-	} else if len(outcome.Records) > 0 {
-		applyRuntimeExecutionRecordProjection(&terminal, id, outcome.Records, finishedAt)
-		projectRuntimeFailure(&terminal.session, &terminal.result, outcome)
-	}
-	applyTerminalRuntimeProjection(state, terminal, outcome)
-	return nil
-}
-
-func TestTaggedDurableHistoryIsAuthoritativeDuringHydrationAndResave(t *testing.T) {
-	t.Parallel()
-	canonical := json.RawMessage(`{"type":"SESSION_STARTED","context":{"sessionId":"dur-sess-tagged"}}`)
-	checkpoint := factory.JavaScriptRuntimeRecord{
-		Sequence: 2,
-		Kind:     factory.JavaScriptRecordKindCheckpoint,
-		Checkpoint: &factory.JavaScriptCheckpointRecord{
-			ID: "checkpoint-tagged", State: map[string]any{"position": "review"},
-		},
-	}
-	mutation := interfaces.TokenMutationRecord{
-		Type: interfaces.MutationMove, TransitionID: "approve", ToPlace: "review:approved",
-	}
-	snapshot := PersistedRuntimeSessionState{
-		Events:         []json.RawMessage{json.RawMessage(`{"type":"LEGACY_EVENT"}`)},
-		RuntimeRecords: []factory.JavaScriptRuntimeRecord{{Sequence: 99, Kind: factory.JavaScriptRecordKindPhase}},
-		Records: []DurableSessionRecord{
-			{Kind: DurableRecordKindCanonicalFactoryEvent, CanonicalEvent: canonical},
-			{Kind: DurableRecordKindJavaScriptRuntime, JavaScriptRecord: &checkpoint},
-			{Kind: DurableRecordKindPetriTokenMutation, PetriMutation: &mutation},
-		},
-	}
-
-	hydrated := runtimeStateFromPersistedSnapshot(snapshot)
-	if len(hydrated.events) != 1 || !bytes.Equal(hydrated.events[0], canonical) {
-		t.Fatalf("hydrated events = %s, want tagged canonical event", hydrated.events)
-	}
-	if len(hydrated.runtimeRecords) != 1 || hydrated.runtimeRecords[0].Checkpoint == nil || hydrated.runtimeRecords[0].Checkpoint.ID != "checkpoint-tagged" {
-		t.Fatalf("hydrated JavaScript records = %#v, want tagged checkpoint", hydrated.runtimeRecords)
-	}
-	if len(hydrated.petriMutations) != 1 || hydrated.petriMutations[0].TransitionID != "approve" || hydrated.petriMutations[0].ToPlace != "review:approved" {
-		t.Fatalf("hydrated Petri mutations = %#v, want tagged transition", hydrated.petriMutations)
-	}
-
-	resaved := persistedSnapshotFromRuntimeState(hydrated)
-	if len(resaved.Records) != 3 || resaved.Records[2].PetriMutation == nil || resaved.Records[2].PetriMutation.TransitionID != "approve" {
-		t.Fatalf("resaved tagged history = %#v, want lossless mixed records", resaved.Records)
-	}
-}
-
-func TestProjectResultRead_ModePartialAndFinal(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-js-run-n-001")
-
-	partial, err := service.GetResult(context.Background(), "dur-sess-js-run-n-001", ResultRequest{Mode: ResultModePartial})
-	if err != nil {
-		t.Fatalf("GetResult partial: %v", err)
-	}
-	if partial.ResultStatus != ResultStatusPartial {
-		t.Fatalf("partial status = %q, want PARTIAL", partial.ResultStatus)
-	}
-	if len(partial.PrimaryResult) == 0 {
-		t.Fatal("partial primaryResult missing")
-	}
-	if partial.Mode != ResultModePartial {
-		t.Fatalf("mode = %q, want partial", partial.Mode)
-	}
-
-	final, err := service.GetResult(context.Background(), "dur-sess-js-run-n-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult final: %v", err)
-	}
-	if final.ResultStatus != ResultStatusNotReady {
-		t.Fatalf("final status = %q, want NOT_READY", final.ResultStatus)
-	}
-	if len(final.PrimaryResult) != 0 {
-		t.Fatal("final primaryResult should be omitted for running session")
-	}
-	if final.Availability == nil || final.Availability.Reason != "RESULT_NOT_READY" {
-		t.Fatalf("availability = %#v, want RESULT_NOT_READY", final.Availability)
-	}
-}
-
-func TestProjectResultRead_TerminalFinalAndUnavailable(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-petri-success-001")
-
-	final, err := service.GetResult(context.Background(), "dur-sess-petri-success-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult terminal final: %v", err)
-	}
-	if final.ResultStatus != ResultStatusFinal {
-		t.Fatalf("status = %q, want FINAL", final.ResultStatus)
-	}
-	if len(final.PrimaryResult) == 0 {
-		t.Fatal("final primaryResult missing")
-	}
-
-	startAsyncByRequestID(t, service, "req-petri-cancel-001")
-	unavailable, err := service.GetResult(context.Background(), "dur-sess-petri-cancel-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult unavailable: %v", err)
-	}
-	if unavailable.ResultStatus != ResultStatusUnavailable {
-		t.Fatalf("status = %q, want UNAVAILABLE", unavailable.ResultStatus)
-	}
-	if unavailable.Availability == nil || unavailable.Availability.Reason != "SESSION_CANCELED" {
-		t.Fatalf("availability = %#v", unavailable.Availability)
-	}
-}
-
-func TestProjectResultRead_FailedWithPartialHonorsPartialMode(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-js-failed-partial-001")
-
-	result, err := service.GetResult(context.Background(), "dur-sess-js-failed-partial-001", ResultRequest{Mode: ResultModePartial})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus != ResultStatusFailedWithPartial {
-		t.Fatalf("status = %q, want FAILED_WITH_PARTIAL", result.ResultStatus)
-	}
-	if len(result.PrimaryResult) == 0 {
-		t.Fatal("partial primaryResult missing")
-	}
-	if result.Failure == nil || !result.Failure.PartialResultAvailable {
-		t.Fatal("failure detail missing")
-	}
-}
-
-func TestProjectResultRead_IncludeArtifactsShaping(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-petri-success-001")
-
-	excluded, err := service.GetResult(context.Background(), "dur-sess-petri-success-001", ResultRequest{
-		Mode:             ResultModeFinal,
-		IncludeArtifacts: false,
-	})
-	if err != nil {
-		t.Fatalf("GetResult excluded: %v", err)
-	}
-	if excluded.IncludeArtifacts {
-		t.Fatal("includeArtifacts = true, want false")
-	}
-	if len(excluded.ArtifactRefs) != 0 {
-		t.Fatalf("artifactRefs = %#v, want omitted", excluded.ArtifactRefs)
-	}
-	if len(excluded.ArtifactIDs) != 1 || excluded.ArtifactIDs[0] != "art-petri-final-001" {
-		t.Fatalf("artifactIds = %#v", excluded.ArtifactIDs)
-	}
-
-	included, err := service.GetResult(context.Background(), "dur-sess-petri-success-001", ResultRequest{
-		Mode:             ResultModeFinal,
-		IncludeArtifacts: true,
-	})
-	if err != nil {
-		t.Fatalf("GetResult included: %v", err)
-	}
-	if !included.IncludeArtifacts {
-		t.Fatal("includeArtifacts = false, want true")
-	}
-	if len(included.ArtifactRefs) != 1 || included.ArtifactRefs[0].ID != "art-petri-final-001" {
-		t.Fatalf("artifactRefs = %#v", included.ArtifactRefs)
-	}
-	if len(included.ArtifactIDs) != 0 {
-		t.Fatalf("artifactIds = %#v, want omitted when refs included", included.ArtifactIDs)
-	}
-}
-
-func TestProjectResultRead_NotReadyRunningSession(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-petri-run-001")
-
-	result, err := service.GetResult(context.Background(), "dur-sess-petri-run-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus != ResultStatusNotReady {
-		t.Fatalf("status = %q, want NOT_READY", result.ResultStatus)
-	}
-	if result.Availability == nil || result.Availability.Message == "" {
-		t.Fatal("availability missing")
-	}
-}
-
-func TestProjectResultRead_DefaultsToFinalMode(t *testing.T) {
-	t.Parallel()
-	canonical := ResultReadResult{
-		SessionID:     "dur-sess-001",
-		ResultStatus:  ResultStatusFinal,
-		SessionStatus: LifecycleStatusSucceeded,
-		PrimaryResult: json.RawMessage(`[{"type":"text","text":"done"}]`),
-	}
-	session := SessionReadResult{
-		SessionID: "dur-sess-001",
-		Status:    LifecycleStatusSucceeded,
-		ResultSummary: &ResultSummary{
-			ResultStatus: string(ResultStatusFinal),
-		},
-	}
-
-	projected, err := ProjectResultRead(canonical, session, nil, ResultRequest{})
-	if err != nil {
-		t.Fatalf("ProjectResultRead: %v", err)
-	}
-	if projected.Mode != ResultModeFinal {
-		t.Fatalf("mode = %q, want final", projected.Mode)
-	}
-	if projected.ResultStatus != ResultStatusFinal {
-		t.Fatalf("status = %q, want FINAL", projected.ResultStatus)
-	}
-}
 
 func TestIsTerminalLifecycleStatus(t *testing.T) {
 	t.Parallel()
@@ -498,1095 +133,6 @@ func TestCheckControlRequestIDReplay_Conflict(t *testing.T) {
 	if !errors.Is(err, ErrControlRequestIDConflict) {
 		t.Fatalf("error = %v, want ErrControlRequestIDConflict", err)
 	}
-}
-
-func TestServiceMethods_PropagateContextCancellation(t *testing.T) {
-	t.Parallel()
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	var service interface {
-		GetSession(context.Context, string) (SessionReadResult, error)
-	}
-	service = stubCancelAwareService{}
-	if _, err := service.GetSession(ctx, "dur-sess-001"); !errors.Is(err, context.Canceled) {
-		t.Fatalf("GetSession error = %v, want context.Canceled", err)
-	}
-}
-
-type stubCancelAwareService struct{}
-
-func (stubCancelAwareService) GetSession(ctx context.Context, _ string) (SessionReadResult, error) {
-	if err := ctx.Err(); err != nil {
-		return SessionReadResult{}, err
-	}
-	return SessionReadResult{}, nil
-}
-
-func (stubCancelAwareService) StartAsync(context.Context, StartRequest) (AsyncStartResult, error) {
-	return AsyncStartResult{}, nil
-}
-func (stubCancelAwareService) StartSync(context.Context, StartRequest) (SyncStartResult, error) {
-	return SyncStartResult{}, nil
-}
-func (stubCancelAwareService) Pause(context.Context, string, ControlRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) Resume(context.Context, string, ControlRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) Cancel(context.Context, string, ControlRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) Terminate(context.Context, string, ControlRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) Approve(context.Context, string, ApproveRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) RetryDispatch(context.Context, string, RetryDispatchRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) InterruptDispatch(context.Context, string, InterruptDispatchRequest) (LifecycleControlResult, error) {
-	return LifecycleControlResult{}, nil
-}
-func (stubCancelAwareService) GetResult(context.Context, string, ResultRequest) (ResultReadResult, error) {
-	return ResultReadResult{}, nil
-}
-func (stubCancelAwareService) ListDispatches(context.Context, string) (ListDispatchesResult, error) {
-	return ListDispatchesResult{}, nil
-}
-func (stubCancelAwareService) GetDispatch(context.Context, string, string) (DispatchDetail, error) {
-	return DispatchDetail{}, nil
-}
-func (stubCancelAwareService) ListArtifacts(context.Context, string) (ListArtifactsResult, error) {
-	return ListArtifactsResult{}, nil
-}
-func (stubCancelAwareService) GetArtifact(context.Context, string, string) (ArtifactDetail, error) {
-	return ArtifactDetail{}, nil
-}
-func (stubCancelAwareService) ReadEvents(context.Context, string, EventReconnectRequest) (EventReadResult, error) {
-	return EventReadResult{}, nil
-}
-
-func (stubCancelAwareService) ListSessions(context.Context, ListSessionsRequest) (ListSessionsResult, error) {
-	return ListSessionsResult{}, nil
-}
-
-const simpleFinalWorkflowSource = `return {
-  label: meta.name,
-  description: meta.description,
-  subject: args.subject,
-  repeat: args.count,
-  echo: args.prefix + ":" + args.subject,
-};
-`
-
-const busyLoopWorkflowSource = `while (true) {}`
-
-const throwErrorWorkflowSource = `throw new Error("workflow execution failed: " + args.subject);`
-
-const progressThenFinalWorkflowSource = `
-phase("execute");
-const artifactRef = workflow.artifact({
-  kind: "log",
-  label: "unpersisted-output",
-  content: { message: "must roll back" },
-});
-workflow.checkpoint({
-  label: "before-final",
-  state: { artifactRef: artifactRef },
-});
-return { artifactRef: artifactRef };
-`
-
-type durableFixedClock struct{ now time.Time }
-
-func (c durableFixedClock) Now() time.Time { return c.now }
-
-func TestJavaScriptRuntimeService_StartSync_UsesInjectedClock(t *testing.T) {
-	t.Parallel()
-	want := time.Date(2031, time.April, 5, 6, 7, 8, 0, time.FixedZone("offset", -7*60*60))
-	projectRoot := t.TempDir()
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: projectRoot,
-		Clock:       durableFixedClock{now: want},
-	})
-
-	started, err := service.StartSync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-clock-001",
-		simpleFinalWorkflowSource,
-		map[string]any{"subject": "clock", "count": 1, "prefix": "fixed"},
-		nil,
-	))
-	if err != nil {
-		t.Fatalf("StartSync: %v", err)
-	}
-	read, err := service.GetSession(context.Background(), started.SessionID)
-	if err != nil {
-		t.Fatalf("GetSession: %v", err)
-	}
-	wantUTC := want.UTC()
-	if read.Lifecycle == nil || read.Lifecycle.StartedAt == nil || !read.Lifecycle.StartedAt.Equal(wantUTC) {
-		t.Fatalf("startedAt = %#v, want %s", read.Lifecycle, wantUTC)
-	}
-	if read.Lifecycle.FinishedAt == nil || !read.Lifecycle.FinishedAt.Equal(wantUTC) {
-		t.Fatalf("finishedAt = %#v, want %s", read.Lifecycle, wantUTC)
-	}
-}
-
-func TestJavaScriptRuntimeService_StartSync_SimpleWorkflowCompletesWithPrimaryResult(t *testing.T) {
-	t.Parallel()
-	service := newDefaultJavaScriptRuntimeService(t, scriptedSuccessfulRuntimeWorkflows(map[string]any{
-		"label":       "runtime-sync-fixture",
-		"description": "runtime sync fixture",
-		"subject":     "workflows",
-		"repeat":      float64(3),
-		"echo":        "you:workflows",
-	}))
-
-	started, err := service.StartSync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-sync-simple-final-001",
-		simpleFinalWorkflowSource,
-		map[string]any{
-			"subject": "workflows",
-			"count":   3,
-			"prefix":  "you",
-		},
-		nil,
-	))
-	if err != nil {
-		t.Fatalf("StartSync: %v", err)
-	}
-	if started.SyncOutcome != SyncOutcomeCompleted {
-		t.Fatalf("syncOutcome = %q, want COMPLETED", started.SyncOutcome)
-	}
-	if started.Status != string(LifecycleStatusSucceeded) {
-		t.Fatalf("status = %q, want SUCCEEDED", started.Status)
-	}
-	if started.OrchestratorKind != interfaces.OrchestratorKindJavaScript {
-		t.Fatalf("orchestratorKind = %q, want JAVASCRIPT", started.OrchestratorKind)
-	}
-	if started.SessionID == "" || started.ResolvedSource.SourceRef == "" || started.SourceHash == "" {
-		t.Fatalf("start result missing resolved source metadata: %#v", started)
-	}
-
-	testJavaScriptRuntimeSyncCompletedSession(t, service, started.SessionID)
-	testJavaScriptRuntimeSyncCompletedResult(t, service, started.SessionID)
-	testJavaScriptRuntimeSyncCompletedEvents(t, service, started.SessionID)
-}
-
-func TestJavaScriptRuntimeService_StartSync_PersistenceFailureDoesNotPublishSuccess(t *testing.T) {
-	t.Parallel()
-	store := &runtimeRecordingStore{saveErr: errors.New("append unavailable")}
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Persistence: store,
-	})
-
-	started, err := service.StartSync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-sync-persist-failure-001",
-		simpleFinalWorkflowSource,
-		nil,
-		nil,
-	))
-	if err == nil || !strings.Contains(err.Error(), "append unavailable") {
-		t.Fatalf("StartSync result = %#v, error = %v, want persistence failure", started, err)
-	}
-
-	store.mu.Lock()
-	saveCalls := store.saveCalls
-	store.mu.Unlock()
-	if saveCalls != 1 {
-		t.Fatalf("persistence save calls = %d, want exactly one", saveCalls)
-	}
-	service.mu.RLock()
-	defer service.mu.RUnlock()
-	for _, state := range service.sessions {
-		if state.session.Status == LifecycleStatusSucceeded || state.result.ResultStatus == ResultStatusFinal {
-			t.Fatalf("unpersisted success became live: session=%#v result=%#v", state.session, state.result)
-		}
-	}
-}
-
-func TestJavaScriptRuntimeService_StartAsync_PersistenceFailurePublishesFailureNotSuccess(t *testing.T) {
-	t.Parallel()
-	store := &runtimeRecordingStore{saveErr: errors.New("append unavailable")}
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Persistence: store,
-	})
-
-	started, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-async-persist-failure-001",
-		progressThenFinalWorkflowSource,
-		nil,
-		nil,
-	))
-	if err != nil {
-		t.Fatalf("StartAsync: %v", err)
-	}
-	failed := waitUntilSessionStatus(t, service, started.SessionID, LifecycleStatusFailed, 2*time.Second)
-	if failed.Failure == nil || !strings.Contains(failed.Failure.Message, "append unavailable") {
-		t.Fatalf("failure = %#v, want explicit persistence failure", failed.Failure)
-	}
-	assertPersistenceFailureRolledBackLiveProjections(t, service, started.SessionID)
-
-	store.mu.Lock()
-	saveCalls := store.saveCalls
-	store.mu.Unlock()
-	if saveCalls != 1 {
-		t.Fatalf("persistence save calls = %d, want exactly one", saveCalls)
-	}
-}
-
-func assertPersistenceFailureRolledBackLiveProjections(
-	t *testing.T,
-	service *JavaScriptRuntimeService,
-	sessionID string,
-) {
-	t.Helper()
-	result, err := service.GetResult(context.Background(), sessionID, ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus == ResultStatusFinal {
-		t.Fatalf("unpersisted terminal result status = %q, want unavailable", result.ResultStatus)
-	}
-	dispatches, err := service.ListDispatches(context.Background(), sessionID)
-	if err != nil {
-		t.Fatalf("ListDispatches: %v", err)
-	}
-	if len(dispatches.Dispatches) != 0 {
-		t.Fatalf("unpersisted dispatches = %#v, want none", dispatches.Dispatches)
-	}
-	artifacts, err := service.ListArtifacts(context.Background(), sessionID)
-	if err != nil {
-		t.Fatalf("ListArtifacts: %v", err)
-	}
-	if len(artifacts.Artifacts) != 0 {
-		t.Fatalf("unpersisted artifacts = %#v, want none", artifacts.Artifacts)
-	}
-	events, err := service.ReadEvents(context.Background(), sessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents: %v", err)
-	}
-	encodedEvents, err := json.Marshal(events.Events)
-	if err != nil {
-		t.Fatalf("marshal events: %v", err)
-	}
-	if bytes.Contains(encodedEvents, []byte("unpersisted-output")) || bytes.Contains(encodedEvents, []byte("checkpoint")) {
-		t.Fatalf("event history retained unpersisted terminal records: %s", encodedEvents)
-	}
-	assertPersistenceFailureClearedInternalRuntimeState(t, service, sessionID)
-}
-
-func assertPersistenceFailureClearedInternalRuntimeState(t *testing.T, service *JavaScriptRuntimeService, sessionID string) {
-	t.Helper()
-	service.mu.RLock()
-	live := cloneRuntimeSessionState(service.sessions[sessionID])
-	service.mu.RUnlock()
-	if live.session.Phase != "" || live.session.Progress != nil || len(live.session.ArtifactRefs) != 0 {
-		t.Fatalf("session retained unpersisted projection: %#v", live.session)
-	}
-	if len(live.runtimeRecords) != 0 || live.checkpointSummary != nil {
-		t.Fatalf("runtime state retained unpersisted records: records=%#v checkpoint=%#v", live.runtimeRecords, live.checkpointSummary)
-	}
-}
-
-type runtimeRecordingStore struct {
-	mu        sync.Mutex
-	saveCalls int
-	saveErr   error
-	payload   []byte
-}
-
-func (s *runtimeRecordingStore) Save(_ string, encoded []byte) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.saveCalls++
-	if s.saveErr != nil {
-		return s.saveErr
-	}
-	s.payload = append([]byte(nil), encoded...)
-	return nil
-}
-
-func (s *runtimeRecordingStore) Load(string) ([]byte, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if len(s.payload) == 0 {
-		return nil, os.ErrNotExist
-	}
-	return append([]byte(nil), s.payload...), nil
-}
-
-func TestJavaScriptRuntimeService_StartAsync_RunningCancelAndReads(t *testing.T) {
-	t.Parallel()
-	service := newDefaultJavaScriptRuntimeService(t, scriptedBlockingRuntimeWorkflows())
-
-	started, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-async-running-001",
-		busyLoopWorkflowSource,
-		map[string]any{"subject": "workflows"},
-		nil,
-	))
-	if err != nil {
-		t.Fatalf("StartAsync: %v", err)
-	}
-	if started.Status != string(LifecycleStatusRunning) {
-		t.Fatalf("start status = %q, want RUNNING", started.Status)
-	}
-
-	session, err := service.GetSession(context.Background(), started.SessionID)
-	if err != nil {
-		t.Fatalf("GetSession: %v", err)
-	}
-	if session.Status != LifecycleStatusRunning {
-		t.Fatalf("session status = %q, want RUNNING", session.Status)
-	}
-
-	result, err := service.GetResult(context.Background(), started.SessionID, ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus != ResultStatusNotReady {
-		t.Fatalf("resultStatus = %q, want NOT_READY", result.ResultStatus)
-	}
-	if result.Availability == nil || result.Availability.Reason != "RESULT_NOT_READY" {
-		t.Fatalf("availability = %#v, want RESULT_NOT_READY", result.Availability)
-	}
-
-	dispatches, err := service.ListDispatches(context.Background(), started.SessionID)
-	if err != nil {
-		t.Fatalf("ListDispatches: %v", err)
-	}
-	if len(dispatches.Dispatches) != 0 {
-		t.Fatalf("dispatches = %#v, want none for busy loop workflow", dispatches.Dispatches)
-	}
-
-	canceled, err := service.Cancel(context.Background(), started.SessionID, ControlRequest{})
-	if err != nil {
-		t.Fatalf("Cancel: %v", err)
-	}
-	if canceled.Outcome != LifecycleControlOutcomeAccepted {
-		t.Fatalf("cancel outcome = %q, want ACCEPTED", canceled.Outcome)
-	}
-
-	finalSession := waitUntilSessionStatus(t, service, started.SessionID, LifecycleStatusCanceled, 5*time.Second)
-	if finalSession.Failure == nil || finalSession.Failure.Reason != "WORKFLOW_RUNTIME_CANCELED" {
-		t.Fatalf("failure = %#v, want WORKFLOW_RUNTIME_CANCELED", finalSession.Failure)
-	}
-}
-
-func TestJavaScriptRuntimeService_StartAsync_FailedAndTimedOut(t *testing.T) {
-	t.Parallel()
-	t.Run("failed", func(t *testing.T) {
-		service := newDefaultJavaScriptRuntimeService(t, scriptedFailedRuntimeWorkflows())
-		started, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-			"req-runtime-async-failed-001",
-			throwErrorWorkflowSource,
-			map[string]any{"subject": "workflows"},
-			nil,
-		))
-		if err != nil {
-			t.Fatalf("StartAsync: %v", err)
-		}
-
-		session := waitUntilSessionStatus(t, service, started.SessionID, LifecycleStatusFailed, 5*time.Second)
-		if session.Failure == nil || session.Failure.Reason == "" {
-			t.Fatalf("failure = %#v, want runtime failure summary", session.Failure)
-		}
-
-		result, err := service.GetResult(context.Background(), started.SessionID, ResultRequest{Mode: ResultModeFinal})
-		if err != nil {
-			t.Fatalf("GetResult: %v", err)
-		}
-		if result.ResultStatus != ResultStatusUnavailable || result.SessionStatus != LifecycleStatusFailed {
-			t.Fatalf("result = %#v, want unavailable failed result", result)
-		}
-	})
-
-	t.Run("timed out", func(t *testing.T) {
-		service := newDefaultJavaScriptRuntimeService(t, scriptedBlockingRuntimeWorkflows())
-		maxRunDurationMs := int64(50)
-		started, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-			"req-runtime-async-timeout-001",
-			busyLoopWorkflowSource,
-			map[string]any{"subject": "workflows"},
-			map[string]any{"maxRunDurationMs": maxRunDurationMs},
-		))
-		if err != nil {
-			t.Fatalf("StartAsync: %v", err)
-		}
-
-		session := waitUntilSessionStatus(t, service, started.SessionID, LifecycleStatusTimedOut, 5*time.Second)
-		if session.Failure == nil || session.Failure.Reason != "WORKFLOW_RUNTIME_TIMEOUT" {
-			t.Fatalf("failure = %#v, want WORKFLOW_RUNTIME_TIMEOUT", session.Failure)
-		}
-
-		result, err := service.GetResult(context.Background(), started.SessionID, ResultRequest{Mode: ResultModeFinal})
-		if err != nil {
-			t.Fatalf("GetResult: %v", err)
-		}
-		if result.ResultStatus != ResultStatusUnavailable || result.SessionStatus != LifecycleStatusTimedOut {
-			t.Fatalf("result = %#v, want unavailable timed out result", result)
-		}
-	})
-}
-
-func TestJavaScriptRuntimeService_StartSync_WaitTimeoutWithoutCancelKeepsSessionRunning(t *testing.T) {
-	t.Parallel()
-	service := newDefaultJavaScriptRuntimeService(t, scriptedBlockingRuntimeWorkflows())
-	waitMillis := int64(50)
-
-	started, err := service.StartSync(context.Background(), StartRequest{
-		RequestID: "req-runtime-sync-wait-timeout-001",
-		Source: Source{
-			Kind: factory.WorkflowSourceKindInlineWorkflow,
-			InlineWorkflow: &InlineWorkflowSource{
-				InlineSource: busyLoopWorkflowSource,
-				Dialect:      "you-workflow-v1",
-				Metadata:     map[string]string{"name": "runtime-sync-wait-fixture"},
-			},
-		},
-		Args: map[string]any{"subject": "workflows"},
-		Wait: &WaitOptions{TimeoutMillis: &waitMillis},
-	})
-	if err != nil {
-		t.Fatalf("StartSync: %v", err)
-	}
-	if started.SyncOutcome != SyncOutcomeTimedOut || !started.TimedOut {
-		t.Fatalf("sync response = %#v, want TIMED_OUT", started)
-	}
-	if started.SessionCanceledByTimeout {
-		t.Fatal("sessionCanceledByTimeout = true, want false")
-	}
-	if started.Status != string(LifecycleStatusRunning) {
-		t.Fatalf("status = %q, want RUNNING", started.Status)
-	}
-
-	session, err := service.GetSession(context.Background(), started.SessionID)
-	if err != nil {
-		t.Fatalf("GetSession: %v", err)
-	}
-	if session.Status != LifecycleStatusRunning {
-		t.Fatalf("session status = %q, want RUNNING after sync wait timeout", session.Status)
-	}
-
-	result, err := service.GetResult(context.Background(), started.SessionID, ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.Availability == nil || result.Availability.Reason != "SYNC_WAIT_TIMED_OUT" {
-		t.Fatalf("availability = %#v, want SYNC_WAIT_TIMED_OUT", result.Availability)
-	}
-}
-
-func TestExecutionServiceAndHelperNormalization(t *testing.T) {
-	t.Parallel()
-	t.Run("execution service providers", testExecutionServiceProviders)
-	t.Run("explicit persistence choices", testExecutionServicePersistenceChoices)
-	t.Run("child executor and smoke provider", testExecutionServiceChildExecutorHelpers)
-	t.Run("source request helpers", testExecutionServiceSourceRequestHelpers)
-}
-
-func testExecutionServiceProviders(t *testing.T) {
-	t.Helper()
-
-	fakeService, err := newExecutionService(ExecutionProviderFake, serviceConfig{})
-	if err != nil {
-		t.Fatalf("NewExecutionService(fake): %v", err)
-	}
-	if _, ok := fakeService.(*FakeService); !ok {
-		t.Fatalf("fake provider type = %T, want *FakeService", fakeService)
-	}
-
-	projectRoot := t.TempDir()
-	persistence, err := ProjectPersistence(projectRoot, testRuntimePersistenceStoreFactory)
-	if err != nil {
-		t.Fatalf("ProjectPersistence: %v", err)
-	}
-	runtimeService, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{ProjectRoot: projectRoot, Persistence: persistence, Clock: durableFixedClock{now: time.Now()}})
-	if err != nil {
-		t.Fatalf("NewExecutionService(runtime): %v", err)
-	}
-	jsService, ok := runtimeService.(*JavaScriptRuntimeService)
-	if !ok {
-		t.Fatalf("runtime provider type = %T, want *JavaScriptRuntimeService", runtimeService)
-	}
-	if jsService.persistence == nil {
-		t.Fatal("expected runtime service to use the injected persisted session store")
-	}
-
-	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{}); err == nil {
-		t.Fatal("NewExecutionService(runtime without projectRoot) error = nil, want validation error")
-	}
-	if _, err := newExecutionService(ExecutionProvider("unknown"), serviceConfig{}); err == nil {
-		t.Fatal("NewExecutionService(unknown) error = nil, want validation error")
-	}
-}
-
-func testExecutionServicePersistenceChoices(t *testing.T) {
-	t.Helper()
-	projectRoot := t.TempDir()
-	testApplicationPersistencePolicies(t, projectRoot)
-	testExecutionServiceRequiredPersistenceDependencies(t, projectRoot)
-	testExecutionServiceDisabledPersistence(t, projectRoot)
-	testExecutionServiceInvalidPersistenceChoices(t, projectRoot)
-}
-
-func testExecutionServiceRequiredPersistenceDependencies(t *testing.T, projectRoot string) {
-	t.Helper()
-	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
-		ProjectRoot: projectRoot,
-		Persistence: DisabledPersistence(),
-	}); err == nil {
-		t.Fatal("NewExecutionService(runtime without clock) error = nil, want validation error")
-	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "clock" {
-		t.Fatalf("runtime without clock error = %#v, want clock ValidationError", err)
-	}
-	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{ProjectRoot: projectRoot}); err == nil {
-		t.Fatal("NewExecutionService(runtime without persistence choice) error = nil, want validation error")
-	}
-}
-
-func testExecutionServiceDisabledPersistence(t *testing.T, projectRoot string) {
-	t.Helper()
-	disabled, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
-		ProjectRoot: projectRoot,
-		Persistence: DisabledPersistence(),
-		Clock:       durableFixedClock{now: time.Now()},
-	})
-	if err != nil {
-		t.Fatalf("NewExecutionService(runtime with disabled persistence): %v", err)
-	}
-	if disabled.(*JavaScriptRuntimeService).persistence != nil {
-		t.Fatal("disabled persistence unexpectedly configured a store")
-	}
-	// A live runtime constructs without a provider on purpose: a runtime-backed
-	// session invokes its children as Workers through a Factory Runtime bound
-	// after construction, so there is nothing a constructor could check. What is
-	// still rejected is a mode this service cannot serve at all.
-	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
-		ProjectRoot:       projectRoot,
-		ChildExecutorMode: ChildExecutorModeLive,
-		Persistence:       DisabledPersistence(),
-		Clock:             durableFixedClock{now: time.Now()},
-	}); err != nil {
-		t.Fatalf("NewExecutionService(live runtime without provider) error = %v, want success", err)
-	}
-	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
-		ProjectRoot:       projectRoot,
-		ChildExecutorMode: "NONSENSE",
-		Persistence:       DisabledPersistence(),
-		Clock:             durableFixedClock{now: time.Now()},
-	}); err == nil {
-		t.Fatal("NewExecutionService(unsupported child executor mode) error = nil, want validation error")
-	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "runtime.childExecutorMode" {
-		t.Fatalf("unsupported child executor mode error = %#v, want runtime.childExecutorMode ValidationError", err)
-	}
-}
-
-func testExecutionServiceInvalidPersistenceChoices(t *testing.T, projectRoot string) {
-	t.Helper()
-	contradictory := PersistenceChoice{store: mustTestRuntimePersistenceStore(t, t.TempDir()), disabled: true}
-	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
-		ProjectRoot: projectRoot,
-		Persistence: contradictory,
-		Clock:       durableFixedClock{now: time.Now()},
-	}); err == nil {
-		t.Fatal("NewExecutionService(runtime with contradictory persistence) error = nil, want validation error")
-	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "persistence" {
-		t.Fatalf("contradictory persistence error = %#v, want persistence ValidationError", err)
-	}
-
-	blockedRoot := filepath.Join(t.TempDir(), "not-a-directory")
-	if err := os.WriteFile(blockedRoot, []byte("blocked"), 0o600); err != nil {
-		t.Fatalf("write blocked persistence root: %v", err)
-	}
-	if _, err := ProjectPersistence(blockedRoot, testRuntimePersistenceStoreFactory); err == nil {
-		t.Fatal("ProjectPersistence(unavailable root) error = nil, want validation error")
-	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "persistence" {
-		t.Fatalf("unavailable persistence error = %#v, want persistence ValidationError", err)
-	}
-}
-
-func testApplicationPersistencePolicies(t *testing.T, projectRoot string) {
-	t.Helper()
-	for _, tc := range []struct {
-		name     string
-		policy   PersistencePolicy
-		disabled bool
-	}{
-		{name: "default disabled", disabled: true},
-		{name: "enabled", policy: PersistencePolicyEnabled},
-		{name: "disabled", policy: PersistencePolicyDisabled, disabled: true},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			choice, err := PersistenceChoiceForPolicy(tc.policy, projectRoot, testRuntimePersistenceStoreFactory)
-			if err != nil {
-				t.Fatalf("PersistenceChoiceForPolicy: %v", err)
-			}
-			store, err := choice.resolve()
-			if err != nil {
-				t.Fatalf("resolve policy choice: %v", err)
-			}
-			if (store == nil) != tc.disabled {
-				t.Fatalf("store nil = %t, want disabled = %t", store == nil, tc.disabled)
-			}
-		})
-	}
-	if _, err := PersistenceChoiceForPolicy(PersistencePolicy("invalid"), projectRoot, testRuntimePersistenceStoreFactory); err == nil {
-		t.Fatal("PersistenceChoiceForPolicy(invalid) error = nil, want validation error")
-	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "persistence.policy" {
-		t.Fatalf("invalid policy error = %#v, want persistence.policy ValidationError", err)
-	}
-}
-
-func TestPersistenceChoiceForPolicy_DefaultDoesNotCreateProjectDurableSessions(t *testing.T) {
-	t.Parallel()
-	projectRoot := t.TempDir()
-	storeCalls := 0
-	choice, err := PersistenceChoiceForPolicy(
-		"",
-		projectRoot,
-		func(string) (runtimepersist.Store, error) {
-			storeCalls++
-			return testRuntimePersistenceStoreFactory(projectRoot)
-		},
-	)
-	if err != nil {
-		t.Fatalf("PersistenceChoiceForPolicy(default): %v", err)
-	}
-	store, err := choice.resolve()
-	if err != nil {
-		t.Fatalf("resolve default policy: %v", err)
-	}
-	if store != nil {
-		t.Fatal("default persistence policy unexpectedly configured a store")
-	}
-	if storeCalls != 0 {
-		t.Fatalf("default policy store factory calls = %d, want 0", storeCalls)
-	}
-	if _, err := os.Stat(runtimepersist.DirForProjectRoot(projectRoot)); !os.IsNotExist(err) {
-		t.Fatalf("durable persistence path stat error = %v, want not exist", err)
-	}
-}
-
-func TestPersistenceChoiceForPolicy_EnabledCreatesProjectDurableSessions(t *testing.T) {
-	t.Parallel()
-	projectRoot := t.TempDir()
-	choice, err := PersistenceChoiceForPolicy(
-		PersistencePolicyEnabled,
-		projectRoot,
-		testRuntimePersistenceStoreFactory,
-	)
-	if err != nil {
-		t.Fatalf("PersistenceChoiceForPolicy(enabled): %v", err)
-	}
-	store, err := choice.resolve()
-	if err != nil {
-		t.Fatalf("resolve enabled policy: %v", err)
-	}
-	if store == nil {
-		t.Fatal("enabled persistence policy returned a nil store")
-	}
-	sessionID := "dur-sess-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-	if err := store.Save(sessionID, []byte(`{"sessionId":"dur-sess-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"}`)); err != nil {
-		t.Fatalf("Save: %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(runtimepersist.DirForProjectRoot(projectRoot), sessionID+".json")); err != nil {
-		t.Fatalf("enabled persistence snapshot stat error = %v, want exist", err)
-	}
-}
-
-func testExecutionServiceChildExecutorHelpers(t *testing.T) {
-	t.Helper()
-
-	if err := validateChildExecutorMode(ChildExecutorModeLive); err != nil {
-		t.Fatalf("validateChildExecutorMode(live) error = %v", err)
-	}
-	if err := validateChildExecutorMode(ChildExecutorModeFake); err != nil {
-		t.Fatalf("validateChildExecutorMode(fake) error = %v", err)
-	}
-	if err := validateChildExecutorMode("nonsense"); err == nil {
-		t.Fatal("validateChildExecutorMode(nonsense) error = nil, want validation error")
-	}
-
-	smoke := SmokeLiveChildProvider()
-	response, err := smoke.Infer(context.Background(), workerexecution.ProviderInferenceRequest{})
-	if err != nil {
-		t.Fatalf("SmokeLiveChildProvider().Infer: %v", err)
-	}
-	if response.ProviderSession == nil || response.ProviderSession.ID == "" {
-		t.Fatalf("provider session = %#v, want stable session metadata", response.ProviderSession)
-	}
-}
-
-func testExecutionServiceSourceRequestHelpers(t *testing.T) {
-	t.Helper()
-
-	inlineReq := startSourceRequest(Source{
-		Kind: factory.WorkflowSourceKindInlineWorkflow,
-		InlineWorkflow: &InlineWorkflowSource{
-			InlineSource: "return 1;",
-		},
-	})
-	if inlineReq.Value != "return 1;" || inlineReq.InlineSource != "return 1;" {
-		t.Fatalf("startSourceRequest(inline) = %#v", inlineReq)
-	}
-	if resolutionOrderForLookupStage(factory.WorkflowSourceLookupStageProjectClaude) != "PROJECT_CLAUDE_WORKFLOWS" {
-		t.Fatal("unexpected lookup stage mapping for project claude")
-	}
-	if resolutionOrderForLookupStage(factory.WorkflowSourceLookupStageNamedJavaScript) != "BUILTIN_GLOBAL_JAVASCRIPT_FACTORIES" {
-		t.Fatal("unexpected lookup stage mapping for named javascript")
-	}
-	if resolutionOrderForLookupStage("unknown") != "" {
-		t.Fatal("unexpected lookup stage mapping for unknown stage")
-	}
-}
-
-func TestNormalizationAndIdempotencyHelpers(t *testing.T) {
-	t.Parallel()
-	t.Run("approve and source normalization", testNormalizationApproveAndSourceBranches)
-	t.Run("canonical json and replay helpers", testNormalizationCanonicalAndReplayBranches)
-	t.Run("idempotency hash stability", testNormalizationIdempotencyHashBranches)
-}
-
-func testNormalizationApproveAndSourceBranches(t *testing.T) {
-	t.Helper()
-
-	approved, err := NormalizeApproveRequest(ApproveRequest{
-		ControlRequest:    ControlRequest{RequestID: "  ctrl-1  ", Reason: "  ok  "},
-		ApprovalPreviewID: "  preview-1  ",
-		ApprovedPolicy:    map[string]any{"policyHash": " hash-1 "},
-	})
-	if err != nil {
-		t.Fatalf("NormalizeApproveRequest: %v", err)
-	}
-	if approved.RequestID != "ctrl-1" || approved.Reason != "ok" || approved.ApprovalPreviewID != "preview-1" {
-		t.Fatalf("normalized approve request = %#v", approved)
-	}
-
-	inlineSource, err := normalizeSourceForIdempotency(Source{
-		Kind: factory.WorkflowSourceKindInlineWorkflow,
-		InlineWorkflow: &InlineWorkflowSource{
-			InlineSource: " return 1; ",
-			Dialect:      " you-workflow-v1 ",
-			Entrypoint:   " default ",
-			Metadata: map[string]string{
-				"b": "2",
-				"a": "1",
-			},
-		},
-	})
-	if err != nil {
-		t.Fatalf("normalizeSourceForIdempotency(inline): %v", err)
-	}
-	inlineWorkflow, ok := inlineSource["inlineWorkflow"].(map[string]any)
-	if !ok || inlineWorkflow["inlineSource"] != "return 1;" {
-		t.Fatalf("inline workflow projection = %#v", inlineSource["inlineWorkflow"])
-	}
-	metadata, ok := inlineWorkflow["metadata"].(map[string]string)
-	if !ok || len(metadata) != 2 || metadata["a"] != "1" || metadata["b"] != "2" {
-		t.Fatalf("inline workflow metadata = %#v", inlineWorkflow["metadata"])
-	}
-
-	if _, err := normalizeSourceForIdempotency(Source{Kind: factory.WorkflowSourceKindInlineWorkflow}); err == nil {
-		t.Fatal("normalizeSourceForIdempotency(missing inline) error = nil, want validation error")
-	}
-}
-
-func testNormalizationCanonicalAndReplayBranches(t *testing.T) {
-	t.Helper()
-
-	if _, err := canonicalizeRawJSON(json.RawMessage("{")); err == nil {
-		t.Fatal("canonicalizeRawJSON(invalid) error = nil, want parse error")
-	}
-	canonical, err := canonicalizeRawJSON(json.RawMessage(`{"b":2,"a":[{"d":4,"c":3}]}`))
-	if err != nil {
-		t.Fatalf("canonicalizeRawJSON(valid): %v", err)
-	}
-	encoded, err := json.Marshal(canonical)
-	if err != nil {
-		t.Fatalf("json.Marshal(canonical): %v", err)
-	}
-	if string(encoded) != `{"a":[{"c":3,"d":4}],"b":2}` {
-		t.Fatalf("canonical json = %s, want sorted object keys", encoded)
-	}
-
-	if err := CheckSyncStartReplayMode(&AsyncStartResult{}, nil, false); !errors.Is(err, ErrExecutionRequestIDConflict) {
-		t.Fatalf("CheckSyncStartReplayMode(async replay mismatch) = %v, want ErrExecutionRequestIDConflict", err)
-	}
-	if err := CheckSyncStartReplayMode(nil, nil, false); err != nil {
-		t.Fatalf("CheckSyncStartReplayMode(empty replay) = %v, want nil", err)
-	}
-	if err := CheckAsyncStartReplayMode(nil); !errors.Is(err, ErrExecutionRequestIDConflict) {
-		t.Fatalf("CheckAsyncStartReplayMode(nil) = %v, want ErrExecutionRequestIDConflict", err)
-	}
-}
-
-func testNormalizationIdempotencyHashBranches(t *testing.T) {
-	t.Helper()
-
-	hashA, err := IdempotencyTupleHash(inlineWorkflowStartRequest("req-hash-1", simpleFinalWorkflowSource, map[string]any{"x": 1}, map[string]any{"policyHash": "same"}))
-	if err != nil {
-		t.Fatalf("IdempotencyTupleHash(first): %v", err)
-	}
-	hashB, err := IdempotencyTupleHash(inlineWorkflowStartRequest("req-hash-1", simpleFinalWorkflowSource, map[string]any{"x": 1}, map[string]any{"policyHash": "same"}))
-	if err != nil {
-		t.Fatalf("IdempotencyTupleHash(second): %v", err)
-	}
-	if hashA != hashB {
-		t.Fatalf("tuple hashes differ: %q vs %q", hashA, hashB)
-	}
-}
-
-func TestPrepareStartAndPersistenceHelpers(t *testing.T) {
-	t.Parallel()
-	projectRoot := writeSimpleFinalWorkflowProject(t)
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: projectRoot,
-		Persistence: mustTestRuntimePersistenceStore(t, runtimepersist.DirForProjectRoot(projectRoot)),
-	})
-
-	prepared, err := service.prepareStart(StartRequest{
-		RequestID: "req-prepare-start-001",
-		Source: Source{
-			Kind:         factory.WorkflowSourceKindWorkflowName,
-			WorkflowName: "simple-final",
-		},
-		Args: map[string]any{
-			"subject": "workflows",
-			"count":   1,
-			"prefix":  "you",
-		},
-	})
-	if err != nil {
-		t.Fatalf("prepareStart: %v", err)
-	}
-	if prepared.SourceRef == "" || prepared.SourceContent == "" || prepared.TupleHash == "" {
-		t.Fatalf("prepared start = %#v, want resolved source fields", prepared)
-	}
-
-	terminal, err := service.executeImmediateSyncSession(context.Background(), prepared.Request, prepared.ResolvedSource, prepared.SourceContent, policyResolutionFromPrepared(prepared), "dur-sess-prepare-001")
-	if err != nil {
-		t.Fatalf("executeImmediateSyncSession: %v", err)
-	}
-	sessionID, err := NewDurableSessionID(testSessionIDGenerator)
-	if err != nil {
-		t.Fatalf("NewDurableSessionID: %v", err)
-	}
-	terminal.session.SessionID = sessionID
-	terminal.result.SessionID = sessionID
-	if terminal.session.Status != LifecycleStatusSucceeded {
-		t.Fatalf("terminal session status = %q, want SUCCEEDED", terminal.session.Status)
-	}
-
-	if err := service.persistTerminalSessionState(terminal); err != nil {
-		t.Fatalf("persistTerminalSessionState: %v", err)
-	}
-	loaded, err := service.snapshotSessionState(sessionID)
-	if err != nil {
-		t.Fatalf("snapshotSessionState(load persisted): %v", err)
-	}
-	if loaded.session.SessionID != sessionID {
-		t.Fatalf("loaded sessionID = %q, want %q", loaded.session.SessionID, sessionID)
-	}
-	if loaded.result.ResultStatus != ResultStatusFinal {
-		t.Fatalf("loaded result status = %q, want FINAL", loaded.result.ResultStatus)
-	}
-}
-
-func TestValidateNamedAgentPresetsRejectsUnknownPresetBeforeStart(t *testing.T) {
-	t.Parallel()
-	err := validateNamedAgentPresets(
-		map[string]interfaces.FactoryOrchestratorJavaScriptAgent{
-			"reviewer": {Preset: "missing-preset"},
-		},
-		map[string]struct{}{"known-preset": {}},
-	)
-	if err == nil || !strings.Contains(err.Error(), `factory agent "reviewer" references unknown operator worker preset "missing-preset"`) {
-		t.Fatalf("validateNamedAgentPresets() error = %v", err)
-	}
-}
-
-func TestJavaScriptRuntimeService_ProjectRootAloneDoesNotEnablePersistence(t *testing.T) {
-	t.Parallel()
-	projectRoot := writeSimpleFinalWorkflowProject(t)
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{ProjectRoot: projectRoot})
-
-	if _, err := service.StartSync(context.Background(), StartRequest{
-		RequestID: "req-runtime-no-implicit-persistence-001",
-		Source: Source{
-			Kind:         factory.WorkflowSourceKindWorkflowName,
-			WorkflowName: "simple-final",
-		},
-	}); err != nil {
-		t.Fatalf("StartSync: %v", err)
-	}
-
-	if _, err := os.Stat(runtimepersist.DirForProjectRoot(projectRoot)); !os.IsNotExist(err) {
-		t.Fatalf("durable persistence path stat error = %v, want not exist", err)
-	}
-}
-
-func newDefaultJavaScriptRuntimeService(t *testing.T, workflows ...factory.JavaScriptWorkflows) *JavaScriptRuntimeService {
-	t.Helper()
-
-	config := javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Clock:       durableFixedClock{now: time.Now()},
-	}
-	if len(workflows) > 0 {
-		config.Workflows = workflows[0]
-	}
-	return newConfiguredJavaScriptRuntimeService(config)
-}
-
-func scriptedRuntimeWorkflows(
-	run func(context.Context, factory.JavaScriptRuntimeRequest, factory.JavaScriptRuntimeHooks) (factory.JavaScriptRuntimeOutcome, error),
-) factory.JavaScriptWorkflows {
-	return factoryruntimefixtures.ScriptedJavaScriptWorkflows{RunFunc: run}
-}
-
-func scriptedSuccessfulRuntimeWorkflows(value map[string]any) factory.JavaScriptWorkflows {
-	return scriptedRuntimeWorkflows(func(
-		context.Context,
-		factory.JavaScriptRuntimeRequest,
-		factory.JavaScriptRuntimeHooks,
-	) (factory.JavaScriptRuntimeOutcome, error) {
-		encoded, err := json.Marshal(value)
-		if err != nil {
-			return factory.JavaScriptRuntimeOutcome{}, err
-		}
-		return factory.JavaScriptRuntimeOutcome{
-			OK:    true,
-			Value: factory.TypedValue{JSON: encoded},
-		}, nil
-	})
-}
-
-func scriptedBlockingRuntimeWorkflows() factory.JavaScriptWorkflows {
-	return scriptedRuntimeWorkflows(func(
-		ctx context.Context,
-		_ factory.JavaScriptRuntimeRequest,
-		_ factory.JavaScriptRuntimeHooks,
-	) (factory.JavaScriptRuntimeOutcome, error) {
-		<-ctx.Done()
-		code := factory.JavaScriptRuntimeCodeCanceled
-		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
-			code = factory.JavaScriptRuntimeCodeTimeout
-		}
-		return factory.JavaScriptRuntimeOutcome{
-			Failure: factory.JavaScriptRuntimeFailure{Code: code, Message: ctx.Err().Error()},
-		}, nil
-	})
-}
-
-func scriptedFailedRuntimeWorkflows() factory.JavaScriptWorkflows {
-	return scriptedRuntimeWorkflows(func(
-		context.Context,
-		factory.JavaScriptRuntimeRequest,
-		factory.JavaScriptRuntimeHooks,
-	) (factory.JavaScriptRuntimeOutcome, error) {
-		return factory.JavaScriptRuntimeOutcome{
-			Failure: factory.JavaScriptRuntimeFailure{
-				Code:    factory.JavaScriptRuntimeCodeScriptError,
-				Message: "scripted workflow execution failure",
-			},
-		}, nil
-	})
-}
-
-func inlineWorkflowStartRequest(
-	requestID string,
-	source string,
-	args map[string]any,
-	requestedPolicy map[string]any,
-) StartRequest {
-	return StartRequest{
-		RequestID: requestID,
-		Source: Source{
-			Kind: factory.WorkflowSourceKindInlineWorkflow,
-			InlineWorkflow: &InlineWorkflowSource{
-				InlineSource: source,
-				Dialect:      "you-workflow-v1",
-				Metadata: map[string]string{
-					"name":        "runtime-async-fixture",
-					"description": "returns a structured final value",
-				},
-			},
-		},
-		Args:            args,
-		RequestedPolicy: requestedPolicy,
-	}
-}
-
-func waitUntilSessionStatus(
-	t *testing.T,
-	service Service,
-	sessionID string,
-	want LifecycleStatus,
-	timeout time.Duration,
-) SessionReadResult {
-	t.Helper()
-
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		session, err := service.GetSession(context.Background(), sessionID)
-		if err != nil {
-			t.Fatalf("GetSession: %v", err)
-		}
-		if session.Status == want {
-			return session
-		}
-		if IsTerminalLifecycleStatus(session.Status) && session.Status != want {
-			t.Fatalf("session %s reached terminal %q before %q", sessionID, session.Status, want)
-		}
-		time.Sleep(10 * time.Millisecond)
-	}
-	t.Fatalf("session %s did not reach status %q within %s", sessionID, want, timeout)
-	return SessionReadResult{}
-}
-
-func decodePrimaryResultMap(t *testing.T, raw json.RawMessage) map[string]any {
-	t.Helper()
-
-	var content []struct {
-		Type string          `json:"type"`
-		JSON json.RawMessage `json:"json,omitempty"`
-	}
-	if err := json.Unmarshal(raw, &content); err != nil {
-		t.Fatalf("unmarshal primary result content: %v", err)
-	}
-	for _, part := range content {
-		if part.Type == "JSON" && len(part.JSON) > 0 {
-			var projected map[string]any
-			if err := json.Unmarshal(part.JSON, &projected); err != nil {
-				t.Fatalf("unmarshal primary result json part: %v", err)
-			}
-			return projected
-		}
-	}
-	t.Fatalf("primary result content = %#v, want JSON part", content)
-	return nil
-}
-
-func writeSimpleFinalWorkflowProject(t *testing.T) string {
-	t.Helper()
-
-	projectRoot := t.TempDir()
-	workflowDir := filepath.Join(projectRoot, factory.WorkflowSourceProjectClaudeWorkflowsDir)
-	if err := os.MkdirAll(workflowDir, 0o755); err != nil {
-		t.Fatalf("mkdir workflows: %v", err)
-	}
-	workflowPath := filepath.Join(workflowDir, "simple-final.workflow.js")
-	if err := os.WriteFile(workflowPath, []byte(simpleFinalWorkflowSource), 0o600); err != nil {
-		t.Fatalf("write workflow: %v", err)
-	}
-	return projectRoot
 }
 
 func TestJavaScriptRuntimeService_ControlWrappersAndDetailReaders(t *testing.T) {
@@ -1734,511 +280,6 @@ func testJavaScriptRuntimeServiceRetryFailedDispatch(t *testing.T, service *Java
 	}
 }
 
-func TestNormalizeStartRequestAndErrorHelpers(t *testing.T) {
-	t.Parallel()
-	t.Run("normalize valid and invalid start requests", testNormalizeStartRequestBranches)
-	t.Run("normalize source and child executor mode", testNormalizeSourceAndExecutorModeBranches)
-	t.Run("error helper strings", testControlAndValidationErrorHelpers)
-}
-
-func testNormalizeStartRequestBranches(t *testing.T) {
-	t.Helper()
-
-	normalized, err := NormalizeStartRequest(StartRequest{
-		RequestID: " req-1 ",
-		Source: Source{
-			Kind:          factory.WorkflowSourceKindFactoryInline,
-			FactoryInline: json.RawMessage(`{"b":2,"a":1}`),
-		},
-		Orchestrator: &OrchestratorOverride{
-			Kind: " custom ",
-			Raw:  json.RawMessage(`{"z":2,"a":1}`),
-		},
-		Runtime: &RuntimeOptions{ChildExecutorMode: " live-provider "},
-	})
-	if err != nil {
-		t.Fatalf("NormalizeStartRequest(factory inline): %v", err)
-	}
-	if normalized.RequestID != "req-1" {
-		t.Fatalf("requestID = %q, want req-1", normalized.RequestID)
-	}
-	if string(normalized.Source.FactoryInline) == "" {
-		t.Fatalf("factory inline unexpectedly empty: %#v", normalized.Source)
-	}
-	if normalized.Runtime == nil || normalized.Runtime.ChildExecutorMode != ChildExecutorModeLive {
-		t.Fatalf("runtime = %#v, want live mode", normalized.Runtime)
-	}
-	if normalized.Orchestrator == nil || normalized.Orchestrator.Kind != "custom" {
-		t.Fatalf("orchestrator = %#v, want trimmed kind", normalized.Orchestrator)
-	}
-
-	if _, err := NormalizeStartRequest(StartRequest{}); err == nil {
-		t.Fatal("NormalizeStartRequest(missing requestID) error = nil, want validation error")
-	}
-	if _, err := NormalizeStartRequest(StartRequest{
-		RequestID: "req-2",
-		Source: Source{
-			Kind:         factory.WorkflowSourceKindWorkflowFile,
-			WorkflowFile: " path/to/workflow.js ",
-		},
-		Orchestrator: &OrchestratorOverride{
-			Raw: json.RawMessage("{"),
-		},
-	}); err == nil {
-		t.Fatal("NormalizeStartRequest(invalid orchestrator) error = nil, want validation error")
-	}
-}
-
-func testNormalizeSourceAndExecutorModeBranches(t *testing.T) {
-	t.Helper()
-
-	if _, err := normalizeSource(Source{Kind: factory.WorkflowSourceKindWorkflowName, WorkflowName: "  "}); err == nil {
-		t.Fatal("normalizeSource(empty workflow name) error = nil, want validation error")
-	}
-	if got := normalizeChildExecutorMode(" live-provider "); got != ChildExecutorModeLive {
-		t.Fatalf("normalizeChildExecutorMode = %q, want live", got)
-	}
-	if got := resolveChildExecutorMode("fake", StartRequest{Runtime: &RuntimeOptions{ChildExecutorMode: "live-provider"}}); got != ChildExecutorModeLive {
-		t.Fatalf("resolveChildExecutorMode = %q, want live override", got)
-	}
-}
-
-func testControlAndValidationErrorHelpers(t *testing.T) {
-	t.Helper()
-
-	var controlErr *ControlError
-	if controlErr.Error() != "" {
-		t.Fatalf("nil control error message = %q, want empty", controlErr.Error())
-	}
-	controlErr = &ControlError{Outcome: LifecycleControlOutcomeConflict}
-	if controlErr.Error() != string(LifecycleControlOutcomeConflict) {
-		t.Fatalf("control error message = %q, want outcome text", controlErr.Error())
-	}
-	var validationErr *ValidationError
-	if validationErr.Error() != "" {
-		t.Fatalf("nil validation error message = %q, want empty", validationErr.Error())
-	}
-}
-
-func TestRuntimeAndValidationHelperBranches(t *testing.T) {
-	t.Parallel()
-	t.Run("child executor hooks and marshal args", testRuntimeHookAndMarshalBranches)
-	t.Run("workflow metadata and source validation errors", testRuntimeMetadataAndSourceValidationBranches)
-	t.Run("policy validation errors", testRuntimePolicyValidationBranches)
-}
-
-func testRuntimeHookAndMarshalBranches(t *testing.T) {
-	t.Helper()
-
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{ProjectRoot: t.TempDir()})
-	if hooks := service.childExecutorHooks(ChildExecutorModeFake, "session-fake"); hooks.NewChildExecutor != nil {
-		t.Fatalf("fake hooks = %#v, want no child executor override", hooks)
-	}
-	liveService := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot:        t.TempDir(),
-		InvocationExecutor: constructorInvocationExecutor{},
-	})
-
-	if hooks := liveService.childExecutorHooks(ChildExecutorModeLive, "session-live"); hooks.NewChildExecutor == nil {
-		t.Fatal("expected live child executor hook")
-	}
-
-	if raw, err := marshalStartArgs(nil); err != nil || raw != nil {
-		t.Fatalf("marshalStartArgs(nil) = %q, %v, want nil,nil", raw, err)
-	}
-	if _, err := marshalStartArgs(map[string]any{"bad": func() {}}); err == nil {
-		t.Fatal("marshalStartArgs(non-json) error = nil, want validation error")
-	}
-}
-
-func testRuntimeMetadataAndSourceValidationBranches(t *testing.T) {
-	t.Helper()
-
-	metadata := workflowMetadataFromResolved(ResolvedSource{
-		SourceRef: "resolved-ref",
-		Metadata: map[string]string{
-			"project": "root",
-		},
-	}, StartRequest{
-		Source: Source{
-			WorkflowName: "named-workflow",
-			InlineWorkflow: &InlineWorkflowSource{
-				Metadata: map[string]string{"team": "ops"},
-			},
-		},
-	})
-	if metadata["name"] != "named-workflow" || metadata["team"] != "ops" || metadata["project"] != "root" {
-		t.Fatalf("workflow metadata = %#v", metadata)
-	}
-
-	if err := validationErrorFromSourceIssues(nil); err == nil || err.Error() == "" {
-		t.Fatalf("validationErrorFromSourceIssues(nil) = %v, want default validation error", err)
-	}
-	if err := validationErrorFromSourceIssues([]factory.WorkflowValidationIssue{{Message: "bad source", Line: 3, Column: 5}}); err == nil || err.Error() != "bad source (line 3, column 5)" {
-		t.Fatalf("validationErrorFromSourceIssues(location) = %v", err)
-	}
-	if err := validationErrorFromSourceIssues([]factory.WorkflowValidationIssue{
-		{Code: factory.WorkflowValidationCodeImportNotFound, Message: "missing module"},
-	}); err == nil || err.Error() != "[workflow.source.notFound] missing module" {
-		t.Fatalf("validationErrorFromSourceIssues(code) = %v", err)
-	}
-	if err := validationErrorFromSourceIssues([]factory.WorkflowValidationIssue{{}}); err == nil || err.Error() != "workflow source validation failed" {
-		t.Fatalf("validationErrorFromSourceIssues(default message) = %v", err)
-	}
-}
-
-func testRuntimePolicyValidationBranches(t *testing.T) {
-	t.Helper()
-
-	if err := validationErrorFromPolicyIssues(nil); err != nil {
-		t.Fatalf("validationErrorFromPolicyIssues(nil) = %v, want nil", err)
-	}
-	if err := validationErrorFromPolicyIssues([]factory.JavaScriptPolicyIssue{{Message: "blocked"}}); err == nil || err.Error() != "blocked" {
-		t.Fatalf("validationErrorFromPolicyIssues = %v, want blocked", err)
-	}
-	if err := validationErrorFromPolicyIssues([]factory.JavaScriptPolicyIssue{{}}); err == nil || err.Error() != "requested policy is invalid" {
-		t.Fatalf("validationErrorFromPolicyIssues(default message) = %v", err)
-	}
-}
-
-func TestStartSourceRequestAndResolutionOrderBranches(t *testing.T) {
-	t.Parallel()
-	cases := []struct {
-		source Source
-		want   string
-	}{
-		{Source{Kind: factory.WorkflowSourceKindFactoryID, FactoryID: "factory-1"}, "factory-1"},
-		{Source{Kind: factory.WorkflowSourceKindFactoryInline, FactoryInline: json.RawMessage(`{"name":"factory"}`)}, `{"name":"factory"}`},
-		{Source{Kind: factory.WorkflowSourceKindWorkflowFile, WorkflowFile: "wf.js"}, "wf.js"},
-		{Source{Kind: factory.WorkflowSourceKindWorkflowName, WorkflowName: "name"}, "name"},
-	}
-	for _, tc := range cases {
-		if got := startSourceRequest(tc.source); got.Value != tc.want {
-			t.Fatalf("startSourceRequest(%s) value = %q, want %q", tc.source.Kind, got.Value, tc.want)
-		}
-	}
-	if got := startSourceRequest(Source{Kind: factory.WorkflowSourceKindInlineWorkflow}); got.Value != "" || got.InlineSource != "" {
-		t.Fatalf("startSourceRequest(missing inline) = %#v, want empty inline request", got)
-	}
-
-	stages := []factory.WorkflowSourceLookupStage{
-		factory.WorkflowSourceLookupStageProjectClaude,
-		factory.WorkflowSourceLookupStageExplicitSourceKind,
-		factory.WorkflowSourceLookupStageGlobalUser,
-		factory.WorkflowSourceLookupStagePackageRelative,
-		factory.WorkflowSourceLookupStageNamedJavaScript,
-		factory.WorkflowSourceLookupStageExplicitFactory,
-	}
-	for _, stage := range stages {
-		if resolutionOrderForLookupStage(stage) == "" {
-			t.Fatalf("resolutionOrderForLookupStage(%q) returned empty mapping", stage)
-		}
-	}
-}
-
-func TestJavaScriptRuntimeService_ReplayAndReadErrorBranches(t *testing.T) {
-	t.Parallel()
-	service := newDefaultJavaScriptRuntimeService(t)
-	req := inlineWorkflowStartRequest(
-		"req-runtime-replay-001",
-		simpleFinalWorkflowSource,
-		map[string]any{"subject": "workflows", "count": 1, "prefix": "you"},
-		nil,
-	)
-
-	first, err := service.StartAsync(context.Background(), req)
-	if err != nil {
-		t.Fatalf("StartAsync(first): %v", err)
-	}
-	second, err := service.StartAsync(context.Background(), req)
-	if err != nil {
-		t.Fatalf("StartAsync(replay): %v", err)
-	}
-	if second.SessionID != first.SessionID {
-		t.Fatalf("replay sessionID = %q, want %q", second.SessionID, first.SessionID)
-	}
-	waitUntilSessionStatus(t, service, first.SessionID, LifecycleStatusSucceeded, 5*time.Second)
-
-	syncReq := inlineWorkflowStartRequest(
-		"req-runtime-replay-sync-001",
-		simpleFinalWorkflowSource,
-		map[string]any{"subject": "workflows", "count": 1, "prefix": "you"},
-		nil,
-	)
-	syncFirst, err := service.StartSync(context.Background(), syncReq)
-	if err != nil {
-		t.Fatalf("StartSync(first): %v", err)
-	}
-	syncSecond, err := service.StartSync(context.Background(), syncReq)
-	if err != nil {
-		t.Fatalf("StartSync(replay): %v", err)
-	}
-	if syncSecond.SessionID != syncFirst.SessionID {
-		t.Fatalf("sync replay sessionID = %q, want %q", syncSecond.SessionID, syncFirst.SessionID)
-	}
-
-	if _, err := service.GetSession(context.Background(), ""); err == nil {
-		t.Fatal("GetSession(empty) error = nil, want validation error")
-	}
-	if _, err := service.GetSession(context.Background(), "dur-sess-dddddddddddddddddddddddddddddddd"); !errors.Is(err, ErrSessionNotFound) {
-		t.Fatalf("GetSession(missing) = %v, want ErrSessionNotFound", err)
-	}
-	if _, err := service.GetDispatch(context.Background(), syncFirst.SessionID, "missing-dispatch"); !errors.Is(err, ErrDispatchNotFound) {
-		t.Fatalf("GetDispatch(missing) = %v, want ErrDispatchNotFound", err)
-	}
-	if _, err := service.GetArtifact(context.Background(), syncFirst.SessionID, "missing-artifact"); !errors.Is(err, ErrArtifactNotFound) {
-		t.Fatalf("GetArtifact(missing) = %v, want ErrArtifactNotFound", err)
-	}
-	if _, err := service.ReadEvents(context.Background(), syncFirst.SessionID, EventReconnectRequest{AfterEventID: "missing"}); !errors.Is(err, ErrReconnectCursorNotFound) {
-		t.Fatalf("ReadEvents(missing cursor) = %v, want ErrReconnectCursorNotFound", err)
-	}
-}
-
-func TestPersistAndMetadataNoOpBranches(t *testing.T) {
-	t.Parallel()
-	if err := (&JavaScriptRuntimeService{}).persistTerminalSessionState(runtimeSessionState{}); err != nil {
-		t.Fatalf("persistTerminalSessionState(no dir) = %v, want nil", err)
-	}
-
-	projectRoot := t.TempDir()
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: projectRoot,
-		Persistence: mustTestRuntimePersistenceStore(t, runtimepersist.DirForProjectRoot(projectRoot)),
-	})
-
-	if err := service.persistTerminalSessionState(runtimeSessionState{
-		session: SessionReadResult{SessionID: "dur-sess-eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee", Status: LifecycleStatusRunning},
-	}); err != nil {
-		t.Fatalf("persistTerminalSessionState(non-terminal) = %v, want nil", err)
-	}
-	if err := service.persistTerminalSessionState(runtimeSessionState{
-		session: SessionReadResult{Status: LifecycleStatusSucceeded},
-	}); err != nil {
-		t.Fatalf("persistTerminalSessionState(empty session id) = %v, want nil", err)
-	}
-
-	metadata := workflowMetadataFromResolved(ResolvedSource{SourceRef: "fallback-ref"}, StartRequest{})
-	if metadata["name"] != "fallback-ref" {
-		t.Fatalf("fallback metadata name = %#v, want fallback-ref", metadata["name"])
-	}
-}
-
-func TestNormalizeStartRequestAdditionalSourceBranches(t *testing.T) {
-	t.Parallel()
-	cases := []StartRequest{
-		{
-			RequestID: "req-file",
-			Source: Source{
-				Kind:         factory.WorkflowSourceKindWorkflowFile,
-				WorkflowFile: " workflow.js ",
-			},
-		},
-		{
-			RequestID: "req-name",
-			Source: Source{
-				Kind:         factory.WorkflowSourceKindWorkflowName,
-				WorkflowName: " named-workflow ",
-			},
-		},
-		{
-			RequestID: "req-inline",
-			Source: Source{
-				Kind: factory.WorkflowSourceKindInlineWorkflow,
-				InlineWorkflow: &InlineWorkflowSource{
-					InlineSource: " return 1; ",
-					Dialect:      " you-workflow-v1 ",
-					Entrypoint:   " default ",
-					Metadata:     map[string]string{"k": "v"},
-				},
-			},
-		},
-	}
-	for _, req := range cases {
-		normalized, err := NormalizeStartRequest(req)
-		if err != nil {
-			t.Fatalf("NormalizeStartRequest(%s): %v", req.RequestID, err)
-		}
-		if normalized.Source.Kind != req.Source.Kind {
-			t.Fatalf("normalized source kind = %q, want %q", normalized.Source.Kind, req.Source.Kind)
-		}
-	}
-	if _, err := normalizeSource(Source{}); err == nil {
-		t.Fatal("normalizeSource(unknown kind) error = nil, want validation error")
-	}
-}
-
-func TestListingFiltersAndNormalizationBranches(t *testing.T) {
-	t.Parallel()
-	now := time.Now().UTC()
-	later := now.Add(2 * time.Hour)
-	summary := DurableSessionListSummary{
-		SessionID:        "dur-sess-filter-1",
-		Status:           LifecycleStatusRunning,
-		OrchestratorKind: "JAVASCRIPT",
-		ResolvedSource: ResolvedSource{
-			Kind:      factory.WorkflowSourceKindWorkflowName,
-			SourceRef: "customer/support",
-			Metadata:  map[string]string{"project": "/workspace/customer"},
-		},
-		Recoverable: true,
-		StaleLease:  true,
-		Lifecycle: &LifecycleTimestamps{
-			QueuedAt:   &now,
-			StartedAt:  &later,
-			UpdatedAt:  &later,
-			FinishedAt: &later,
-		},
-	}
-	yes := true
-	after := now.Add(-time.Minute)
-	before := later.Add(time.Minute)
-	if !MatchesDurableSessionListFilters(summary, SessionListFilters{
-		Statuses:          []LifecycleStatus{LifecycleStatusRunning},
-		OrchestratorKinds: []string{" javascript "},
-		SourceKind:        factory.WorkflowSourceKindWorkflowName,
-		SourceRef:         "support",
-		ProjectBoundary:   "workspace",
-		Recoverable:       &yes,
-		StaleLease:        &yes,
-		CreatedAfter:      &after,
-		CreatedBefore:     &before,
-		UpdatedAfter:      &after,
-		UpdatedBefore:     &before,
-	}) {
-		t.Fatal("expected summary to match all listing filters")
-	}
-	no := false
-	if MatchesDurableSessionListFilters(summary, SessionListFilters{Recoverable: &no}) {
-		t.Fatal("recoverable mismatch unexpectedly matched")
-	}
-	if containsLifecycleStatus([]LifecycleStatus{LifecycleStatusPaused}, LifecycleStatusRunning) {
-		t.Fatal("containsLifecycleStatus mismatch unexpectedly matched")
-	}
-	if containsString([]string{"Alpha"}, "beta") {
-		t.Fatal("containsString mismatch unexpectedly matched")
-	}
-	if firstLifecycleTimestamp(nil, &later) != &later {
-		t.Fatal("firstLifecycleTimestamp did not return first non-nil value")
-	}
-	if latestLifecycleTimestamp(summary.Lifecycle) != &later {
-		t.Fatal("latestLifecycleTimestamp did not return latest time")
-	}
-
-	normalized, err := NormalizeListSessionsRequest(ListSessionsRequest{
-		Scope: SessionListScopeAll,
-		Filters: SessionListFilters{
-			Statuses:          []LifecycleStatus{LifecycleStatusRunning},
-			OrchestratorKinds: []string{" JAVASCRIPT ", ""},
-			SourceKind:        factory.WorkflowSourceKindWorkflowName,
-			CreatedAfter:      &after,
-			CreatedBefore:     &before,
-		},
-	})
-	if err != nil {
-		t.Fatalf("NormalizeListSessionsRequest: %v", err)
-	}
-	if normalized.Scope != SessionListScopeAll || len(normalized.Filters.OrchestratorKinds) != 1 {
-		t.Fatalf("normalized list request = %#v", normalized)
-	}
-	if _, err := NormalizeListSessionsRequest(ListSessionsRequest{Scope: SessionListScope("bad")}); err == nil {
-		t.Fatal("NormalizeListSessionsRequest(bad scope) error = nil, want validation error")
-	}
-	if _, err := NormalizeListSessionsRequest(ListSessionsRequest{
-		Filters: SessionListFilters{
-			SourceKind:    factory.WorkflowSourceKind("unknown"),
-			CreatedAfter:  &before,
-			CreatedBefore: &after,
-		},
-	}); err == nil {
-		t.Fatal("NormalizeListSessionsRequest(invalid filters) error = nil, want validation error")
-	}
-}
-
-func TestSmallHelperBranches(t *testing.T) {
-	t.Parallel()
-	if got := resolvedDialect(ResolvedSource{Dialect: "custom"}); got != "custom" {
-		t.Fatalf("resolvedDialect(custom) = %q, want custom", got)
-	}
-	if got := resolvedDialect(ResolvedSource{}); got != "you-workflow-v1" {
-		t.Fatalf("resolvedDialect(default) = %q, want you-workflow-v1", got)
-	}
-	if id, err := NormalizeSessionID(" session-1 "); err != nil || id != "session-1" {
-		t.Fatalf("NormalizeSessionID = %q, %v, want session-1,nil", id, err)
-	}
-	if _, err := NormalizeSessionID("   "); err == nil {
-		t.Fatal("NormalizeSessionID(empty) error = nil, want validation error")
-	}
-}
-
-func TestProjectionCloneHelpers(t *testing.T) {
-	t.Parallel()
-	observedAt := time.Now().UTC()
-	artifact := artifactSummaryFromRuntimeRecord("dur-sess-helper-1", factory.JavaScriptArtifactRecord{
-		ID:         "art-helper-1",
-		Kind:       "RESULT",
-		Visibility: "PUBLIC",
-		Label:      "helper",
-	}, observedAt)
-	if artifact.ID != "art-helper-1" || artifact.RetrievalRef == nil || artifact.RetrievalRef.Href == "" {
-		t.Fatalf("artifact summary = %#v", artifact)
-	}
-
-	js := cloneDispatchJavaScriptProjections(map[string]DispatchJavaScriptProjection{
-		"disp-1": {TaskLabel: "child"},
-	})
-	if js["disp-1"].TaskLabel != "child" {
-		t.Fatalf("cloned javascript projections = %#v", js)
-	}
-	transitions := cloneDispatchStatusTransitions(map[string][]DispatchStatus{
-		"disp-1": {DispatchStatusQueued, DispatchStatusRunning},
-	})
-	if len(transitions["disp-1"]) != 2 {
-		t.Fatalf("cloned transitions = %#v", transitions)
-	}
-}
-
-func testJavaScriptRuntimeSyncCompletedSession(t *testing.T, service *JavaScriptRuntimeService, sessionID string) {
-	t.Helper()
-
-	session, err := service.GetSession(context.Background(), sessionID)
-	if err != nil {
-		t.Fatalf("GetSession: %v", err)
-	}
-	if session.Status != LifecycleStatusSucceeded {
-		t.Fatalf("session status = %q, want SUCCEEDED", session.Status)
-	}
-	if session.ResultSummary == nil || session.ResultSummary.ResultStatus != string(ResultStatusFinal) {
-		t.Fatalf("resultSummary = %#v, want FINAL", session.ResultSummary)
-	}
-}
-
-func testJavaScriptRuntimeSyncCompletedResult(t *testing.T, service *JavaScriptRuntimeService, sessionID string) {
-	t.Helper()
-
-	result, err := service.GetResult(context.Background(), sessionID, ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus != ResultStatusFinal {
-		t.Fatalf("resultStatus = %q, want FINAL", result.ResultStatus)
-	}
-	projected := decodePrimaryResultMap(t, result.PrimaryResult)
-	if projected["echo"] != "you:workflows" {
-		t.Fatalf("primaryResult echo = %#v, want you:workflows", projected["echo"])
-	}
-}
-
-func testJavaScriptRuntimeSyncCompletedEvents(t *testing.T, service *JavaScriptRuntimeService, sessionID string) {
-	t.Helper()
-
-	events, err := service.ReadEvents(context.Background(), sessionID, EventReconnectRequest{})
-	if err != nil {
-		t.Fatalf("ReadEvents: %v", err)
-	}
-	if len(events.Events) != 3 {
-		t.Fatalf("events = %d, want 3 canonical lifecycle events", len(events.Events))
-	}
-}
 func TestProjectedLifecycleControlStatus_PrefersCanonicalBracketStatus(t *testing.T) {
 	t.Parallel()
 	status := ProjectedLifecycleControlStatus("PAUSED", "RUNNING")
@@ -2287,6 +328,7 @@ func TestLiveLifecycleControlResponse_BuildsTypedPauseOutcome(t *testing.T) {
 		t.Fatalf("links = %#v, want /factory-sessions/~default", result.Links)
 	}
 }
+
 func TestJavaScriptRuntimeService_InterruptAcceptedBeforeChildCompletion_RecordsObservedCancellation(t *testing.T) {
 	t.Parallel()
 	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{ProjectRoot: t.TempDir()})
@@ -3074,34 +1116,6 @@ func TestResumeHelperFunctions_CoverMergeCloneAndPolicyPaths(t *testing.T) {
 	}
 }
 
-func TestApplyRuntimeSuccessProjection_InvalidResultMarksFailed(t *testing.T) {
-	t.Parallel()
-	sessionID := "dur-sess-invalid-result-001"
-	foreignURI := factory.FormatArtifactURI("dur-sess-other-001", "artifact-1")
-	raw, err := json.Marshal(foreignURI)
-	if err != nil {
-		t.Fatalf("marshal foreign uri: %v", err)
-	}
-	state := &runtimeSessionState{
-		artifacts: []ArtifactSummary{{
-			ID:         "artifact-1",
-			Kind:       "IMAGE",
-			Label:      "output",
-			Visibility: "PUBLIC",
-		}},
-	}
-	applyRuntimeSuccessProjection(state, sessionID, factory.JavaScriptRuntimeOutcome{
-		OK:    true,
-		Value: factory.TypedValue{JSON: raw},
-	}, time.Now().UTC())
-	if state.session.Status != LifecycleStatusFailed {
-		t.Fatalf("status = %q, want FAILED", state.session.Status)
-	}
-	if state.session.Failure == nil || state.session.Failure.Reason != "WORKFLOW_RUNTIME_INVALID_RESULT" {
-		t.Fatalf("failure = %#v, want WORKFLOW_RUNTIME_INVALID_RESULT", state.session.Failure)
-	}
-}
-
 func TestCheckpointEventProjection_BuildsCanonicalCheckpointEvents(t *testing.T) {
 	t.Parallel()
 	startedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
@@ -3629,148 +1643,4 @@ func TestJavaScriptRuntimeServiceWriteRecordingUsesCanonicalSnapshotAndCorrelate
 	if readErr != nil || read.Status != LifecycleStatusSucceeded {
 		t.Fatalf("live session changed after recording failure: read=%#v err=%v", read, readErr)
 	}
-}
-
-func TestJavaScriptRuntimeService_StartSync_WorkflowFilePolicyDeniesDisallowedModel(t *testing.T) {
-	t.Parallel()
-
-	projectRoot := t.TempDir()
-	workflowPath := filepath.Join(projectRoot, "workflow.js")
-	workflowSource := `agent.run({
-  prompt: "summarize workflows",
-  label: "denied-model",
-  model: "gpt-denied",
-});
-return { ok: true };`
-	if err := os.WriteFile(workflowPath, []byte(workflowSource), 0o600); err != nil {
-		t.Fatalf("write workflow: %v", err)
-	}
-
-	defaultPolicy := json.RawMessage(`{
-  "mode":"READ_ONLY",
-  "maxAgents":4,
-  "concurrency":2,
-  "allowedModels":["gpt-allowed"],
-  "allowedReasoningEfforts":["low"]
-}`)
-	var requestedPolicy map[string]any
-	if err := json.Unmarshal(defaultPolicy, &requestedPolicy); err != nil {
-		t.Fatalf("unmarshal default policy: %v", err)
-	}
-
-	workflows := policyDeniedModelWorkflows(workflowSource)
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot:       projectRoot,
-		ChildExecutorMode: ChildExecutorModeFake,
-		Workflows:         workflows,
-	})
-
-	started, err := service.StartSync(context.Background(), StartRequest{
-		RequestID: "req-policy-denied-model",
-		Source: Source{
-			Kind:         factory.WorkflowSourceKindWorkflowFile,
-			WorkflowFile: workflowPath,
-			InlineWorkflow: &InlineWorkflowSource{
-				DefaultPolicy: defaultPolicy,
-			},
-		},
-		Args:            map[string]any{"prompt": "hello"},
-		RequestedPolicy: requestedPolicy,
-		Runtime:         &RuntimeOptions{ChildExecutorMode: ChildExecutorModeFake},
-	})
-	if err != nil {
-		t.Fatalf("StartSync: %v", err)
-	}
-	if started.Status != string(LifecycleStatusFailed) {
-		t.Fatalf("status = %q, want FAILED; outcome = %#v", started.Status, started)
-	}
-
-	session, err := service.GetSession(context.Background(), started.SessionID)
-	if err != nil {
-		t.Fatalf("GetSession: %v", err)
-	}
-	result, err := service.GetResult(context.Background(), started.SessionID, ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	failureMessage := ""
-	if result.Failure != nil {
-		failureMessage = result.Failure.Message
-	} else if session.Failure != nil {
-		failureMessage = session.Failure.Message
-	}
-	if !strings.Contains(failureMessage, `policy denied: model "gpt-denied" is not listed in allowedModels`) {
-		t.Fatalf("session failure = %#v result failure = %#v, want stable policy diagnostic", session.Failure, result.Failure)
-	}
-}
-
-func policyDeniedModelWorkflows(workflowSource string) factory.JavaScriptWorkflows {
-	policyMessage := `policy denied: model "gpt-denied" is not listed in allowedModels`
-	return factoryruntimefixtures.ScriptedJavaScriptWorkflows{
-		ResolveSourceFunc: func(
-			request factory.WorkflowSourceRequest,
-			_ factory.WorkflowSourceContext,
-		) factory.WorkflowSourceResolution {
-			return factory.WorkflowSourceResolution{
-				RequestKind:  request.Kind,
-				RequestValue: request.Value,
-				ResolvedKind: request.Kind,
-				SourceRef:    request.Value,
-				SourceHash:   "sha256:policy-denied",
-				Dialect:      "you-workflow-v1",
-				Content:      workflowSource,
-				Found:        true,
-			}
-		},
-		LoadSourceFunc: func(request factory.WorkflowValidationLoadRequest) (
-			factory.WorkflowValidationLoadedSource,
-			[]factory.WorkflowValidationIssue,
-		) {
-			return factory.WorkflowValidationLoadedSource{
-				SourceRef:        request.SourceRef,
-				SourceHash:       "sha256:policy-denied",
-				Format:           factory.WorkflowValidationFormatJavaScript,
-				AuthoredSource:   request.Content,
-				ExecutableSource: request.Content,
-			}, nil
-		},
-		RunFunc: func(
-			context.Context,
-			factory.JavaScriptRuntimeRequest,
-			factory.JavaScriptRuntimeHooks,
-		) (factory.JavaScriptRuntimeOutcome, error) {
-			return factory.JavaScriptRuntimeOutcome{
-				Failure: factory.JavaScriptRuntimeFailure{
-					Code:    factory.JavaScriptRuntimeCodeScriptError,
-					Message: policyMessage,
-				},
-			}, nil
-		},
-	}
-}
-
-type orchestrationJavaScriptAdapter struct {
-	factory.JavaScriptWorkflowRuntime
-}
-
-func orchestrationJavaScriptFromWorkflows(workflows factory.JavaScriptWorkflows) factory.OrchestrationJavaScriptExecution {
-	if workflows == nil {
-		return nil
-	}
-	return orchestrationJavaScriptAdapter{workflows}
-}
-
-func (a orchestrationJavaScriptAdapter) RunJavaScript(
-	ctx context.Context,
-	req factory.JavaScriptRuntimeRequest,
-	hooks factory.JavaScriptRuntimeHooks,
-) (factory.JavaScriptRuntimeOutcome, error) {
-	return a.Run(ctx, req, hooks)
-}
-
-func (a orchestrationJavaScriptAdapter) ResumeJavaScript(
-	summary factory.JavaScriptCompletedCheckpointSummary,
-	records []factory.JavaScriptRuntimeRecord,
-) factory.JavaScriptResumeContext {
-	return a.ResumeContext(summary, records)
 }
