+++ pkg/services/factory_sessions/internal/execution/javascript_runtime_persistence_test.go
@@ -1,0 +1,482 @@
+package factorysessionexecution
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"errors"
+	"os"
+	"path/filepath"
+	"strings"
+	"sync"
+	"testing"
+	"time"
+
+	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	"github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/execution/runtimepersist"
+)
+
+func TestTaggedDurableHistoryIsAuthoritativeDuringHydrationAndResave(t *testing.T) {
+	t.Parallel()
+	canonical := json.RawMessage(`{"type":"SESSION_STARTED","context":{"sessionId":"dur-sess-tagged"}}`)
+	checkpoint := factory.JavaScriptRuntimeRecord{
+		Sequence: 2,
+		Kind:     factory.JavaScriptRecordKindCheckpoint,
+		Checkpoint: &factory.JavaScriptCheckpointRecord{
+			ID: "checkpoint-tagged", State: map[string]any{"position": "review"},
+		},
+	}
+	mutation := interfaces.TokenMutationRecord{
+		Type: interfaces.MutationMove, TransitionID: "approve", ToPlace: "review:approved",
+	}
+	snapshot := PersistedRuntimeSessionState{
+		Events:         []json.RawMessage{json.RawMessage(`{"type":"LEGACY_EVENT"}`)},
+		RuntimeRecords: []factory.JavaScriptRuntimeRecord{{Sequence: 99, Kind: factory.JavaScriptRecordKindPhase}},
+		Records: []DurableSessionRecord{
+			{Kind: DurableRecordKindCanonicalFactoryEvent, CanonicalEvent: canonical},
+			{Kind: DurableRecordKindJavaScriptRuntime, JavaScriptRecord: &checkpoint},
+			{Kind: DurableRecordKindPetriTokenMutation, PetriMutation: &mutation},
+		},
+	}
+
+	hydrated := runtimeStateFromPersistedSnapshot(snapshot)
+	if len(hydrated.events) != 1 || !bytes.Equal(hydrated.events[0], canonical) {
+		t.Fatalf("hydrated events = %s, want tagged canonical event", hydrated.events)
+	}
+	if len(hydrated.runtimeRecords) != 1 || hydrated.runtimeRecords[0].Checkpoint == nil || hydrated.runtimeRecords[0].Checkpoint.ID != "checkpoint-tagged" {
+		t.Fatalf("hydrated JavaScript records = %#v, want tagged checkpoint", hydrated.runtimeRecords)
+	}
+	if len(hydrated.petriMutations) != 1 || hydrated.petriMutations[0].TransitionID != "approve" || hydrated.petriMutations[0].ToPlace != "review:approved" {
+		t.Fatalf("hydrated Petri mutations = %#v, want tagged transition", hydrated.petriMutations)
+	}
+
+	resaved := persistedSnapshotFromRuntimeState(hydrated)
+	if len(resaved.Records) != 3 || resaved.Records[2].PetriMutation == nil || resaved.Records[2].PetriMutation.TransitionID != "approve" {
+		t.Fatalf("resaved tagged history = %#v, want lossless mixed records", resaved.Records)
+	}
+}
+
+func TestJavaScriptRuntimeService_StartSync_PersistenceFailureDoesNotPublishSuccess(t *testing.T) {
+	t.Parallel()
+	store := &runtimeRecordingStore{saveErr: errors.New("append unavailable")}
+	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
+		ProjectRoot: t.TempDir(),
+		Persistence: store,
+	})
+
+	started, err := service.StartSync(context.Background(), inlineWorkflowStartRequest(
+		"req-runtime-sync-persist-failure-001",
+		simpleFinalWorkflowSource,
+		nil,
+		nil,
+	))
+	if err == nil || !strings.Contains(err.Error(), "append unavailable") {
+		t.Fatalf("StartSync result = %#v, error = %v, want persistence failure", started, err)
+	}
+
+	store.mu.Lock()
+	saveCalls := store.saveCalls
+	store.mu.Unlock()
+	if saveCalls != 1 {
+		t.Fatalf("persistence save calls = %d, want exactly one", saveCalls)
+	}
+	service.mu.RLock()
+	defer service.mu.RUnlock()
+	for _, state := range service.sessions {
+		if state.session.Status == LifecycleStatusSucceeded || state.result.ResultStatus == ResultStatusFinal {
+			t.Fatalf("unpersisted success became live: session=%#v result=%#v", state.session, state.result)
+		}
+	}
+}
+
+func TestJavaScriptRuntimeService_StartAsync_PersistenceFailurePublishesFailureNotSuccess(t *testing.T) {
+	t.Parallel()
+	store := &runtimeRecordingStore{saveErr: errors.New("append unavailable")}
+	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
+		ProjectRoot: t.TempDir(),
+		Persistence: store,
+	})
+
+	started, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
+		"req-runtime-async-persist-failure-001",
+		progressThenFinalWorkflowSource,
+		nil,
+		nil,
+	))
+	if err != nil {
+		t.Fatalf("StartAsync: %v", err)
+	}
+	failed := waitUntilSessionStatus(t, service, started.SessionID, LifecycleStatusFailed, 2*time.Second)
+	if failed.Failure == nil || !strings.Contains(failed.Failure.Message, "append unavailable") {
+		t.Fatalf("failure = %#v, want explicit persistence failure", failed.Failure)
+	}
+	assertPersistenceFailureRolledBackLiveProjections(t, service, started.SessionID)
+
+	store.mu.Lock()
+	saveCalls := store.saveCalls
+	store.mu.Unlock()
+	if saveCalls != 1 {
+		t.Fatalf("persistence save calls = %d, want exactly one", saveCalls)
+	}
+}
+
+func assertPersistenceFailureRolledBackLiveProjections(
+	t *testing.T,
+	service *JavaScriptRuntimeService,
+	sessionID string,
+) {
+	t.Helper()
+	result, err := service.GetResult(context.Background(), sessionID, ResultRequest{Mode: ResultModeFinal})
+	if err != nil {
+		t.Fatalf("GetResult: %v", err)
+	}
+	if result.ResultStatus == ResultStatusFinal {
+		t.Fatalf("unpersisted terminal result status = %q, want unavailable", result.ResultStatus)
+	}
+	dispatches, err := service.ListDispatches(context.Background(), sessionID)
+	if err != nil {
+		t.Fatalf("ListDispatches: %v", err)
+	}
+	if len(dispatches.Dispatches) != 0 {
+		t.Fatalf("unpersisted dispatches = %#v, want none", dispatches.Dispatches)
+	}
+	artifacts, err := service.ListArtifacts(context.Background(), sessionID)
+	if err != nil {
+		t.Fatalf("ListArtifacts: %v", err)
+	}
+	if len(artifacts.Artifacts) != 0 {
+		t.Fatalf("unpersisted artifacts = %#v, want none", artifacts.Artifacts)
+	}
+	events, err := service.ReadEvents(context.Background(), sessionID, EventReconnectRequest{})
+	if err != nil {
+		t.Fatalf("ReadEvents: %v", err)
+	}
+	encodedEvents, err := json.Marshal(events.Events)
+	if err != nil {
+		t.Fatalf("marshal events: %v", err)
+	}
+	if bytes.Contains(encodedEvents, []byte("unpersisted-output")) || bytes.Contains(encodedEvents, []byte("checkpoint")) {
+		t.Fatalf("event history retained unpersisted terminal records: %s", encodedEvents)
+	}
+	assertPersistenceFailureClearedInternalRuntimeState(t, service, sessionID)
+}
+
+func assertPersistenceFailureClearedInternalRuntimeState(t *testing.T, service *JavaScriptRuntimeService, sessionID string) {
+	t.Helper()
+	service.mu.RLock()
+	live := cloneRuntimeSessionState(service.sessions[sessionID])
+	service.mu.RUnlock()
+	if live.session.Phase != "" || live.session.Progress != nil || len(live.session.ArtifactRefs) != 0 {
+		t.Fatalf("session retained unpersisted projection: %#v", live.session)
+	}
+	if len(live.runtimeRecords) != 0 || live.checkpointSummary != nil {
+		t.Fatalf("runtime state retained unpersisted records: records=%#v checkpoint=%#v", live.runtimeRecords, live.checkpointSummary)
+	}
+}
+
+type runtimeRecordingStore struct {
+	mu        sync.Mutex
+	saveCalls int
+	saveErr   error
+	payload   []byte
+}
+
+func (s *runtimeRecordingStore) Save(_ string, encoded []byte) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	s.saveCalls++
+	if s.saveErr != nil {
+		return s.saveErr
+	}
+	s.payload = append([]byte(nil), encoded...)
+	return nil
+}
+
+func (s *runtimeRecordingStore) Load(string) ([]byte, error) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	if len(s.payload) == 0 {
+		return nil, os.ErrNotExist
+	}
+	return append([]byte(nil), s.payload...), nil
+}
+
+func testExecutionServicePersistenceChoices(t *testing.T) {
+	t.Helper()
+	projectRoot := t.TempDir()
+	testApplicationPersistencePolicies(t, projectRoot)
+	testExecutionServiceRequiredPersistenceDependencies(t, projectRoot)
+	testExecutionServiceDisabledPersistence(t, projectRoot)
+	testExecutionServiceInvalidPersistenceChoices(t, projectRoot)
+}
+
+func testExecutionServiceRequiredPersistenceDependencies(t *testing.T, projectRoot string) {
+	t.Helper()
+	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
+		ProjectRoot: projectRoot,
+		Persistence: DisabledPersistence(),
+	}); err == nil {
+		t.Fatal("NewExecutionService(runtime without clock) error = nil, want validation error")
+	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "clock" {
+		t.Fatalf("runtime without clock error = %#v, want clock ValidationError", err)
+	}
+	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{ProjectRoot: projectRoot}); err == nil {
+		t.Fatal("NewExecutionService(runtime without persistence choice) error = nil, want validation error")
+	}
+}
+
+func testExecutionServiceDisabledPersistence(t *testing.T, projectRoot string) {
+	t.Helper()
+	disabled, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
+		ProjectRoot: projectRoot,
+		Persistence: DisabledPersistence(),
+		Clock:       durableFixedClock{now: time.Now()},
+	})
+	if err != nil {
+		t.Fatalf("NewExecutionService(runtime with disabled persistence): %v", err)
+	}
+	if disabled.(*JavaScriptRuntimeService).persistence != nil {
+		t.Fatal("disabled persistence unexpectedly configured a store")
+	}
+	// A live runtime constructs without a provider on purpose: a runtime-backed
+	// session invokes its children as Workers through a Factory Runtime bound
+	// after construction, so there is nothing a constructor could check. What is
+	// still rejected is a mode this service cannot serve at all.
+	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
+		ProjectRoot:       projectRoot,
+		ChildExecutorMode: ChildExecutorModeLive,
+		Persistence:       DisabledPersistence(),
+		Clock:             durableFixedClock{now: time.Now()},
+	}); err != nil {
+		t.Fatalf("NewExecutionService(live runtime without provider) error = %v, want success", err)
+	}
+	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
+		ProjectRoot:       projectRoot,
+		ChildExecutorMode: "NONSENSE",
+		Persistence:       DisabledPersistence(),
+		Clock:             durableFixedClock{now: time.Now()},
+	}); err == nil {
+		t.Fatal("NewExecutionService(unsupported child executor mode) error = nil, want validation error")
+	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "runtime.childExecutorMode" {
+		t.Fatalf("unsupported child executor mode error = %#v, want runtime.childExecutorMode ValidationError", err)
+	}
+}
+
+func testExecutionServiceInvalidPersistenceChoices(t *testing.T, projectRoot string) {
+	t.Helper()
+	contradictory := PersistenceChoice{store: mustTestRuntimePersistenceStore(t, t.TempDir()), disabled: true}
+	if _, err := newExecutionService(ExecutionProviderJavaScriptRuntime, serviceConfig{
+		ProjectRoot: projectRoot,
+		Persistence: contradictory,
+		Clock:       durableFixedClock{now: time.Now()},
+	}); err == nil {
+		t.Fatal("NewExecutionService(runtime with contradictory persistence) error = nil, want validation error")
+	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "persistence" {
+		t.Fatalf("contradictory persistence error = %#v, want persistence ValidationError", err)
+	}
+
+	blockedRoot := filepath.Join(t.TempDir(), "not-a-directory")
+	if err := os.WriteFile(blockedRoot, []byte("blocked"), 0o600); err != nil {
+		t.Fatalf("write blocked persistence root: %v", err)
+	}
+	if _, err := ProjectPersistence(blockedRoot, testRuntimePersistenceStoreFactory); err == nil {
+		t.Fatal("ProjectPersistence(unavailable root) error = nil, want validation error")
+	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "persistence" {
+		t.Fatalf("unavailable persistence error = %#v, want persistence ValidationError", err)
+	}
+}
+
+func testApplicationPersistencePolicies(t *testing.T, projectRoot string) {
+	t.Helper()
+	for _, tc := range []struct {
+		name     string
+		policy   PersistencePolicy
+		disabled bool
+	}{
+		{name: "default disabled", disabled: true},
+		{name: "enabled", policy: PersistencePolicyEnabled},
+		{name: "disabled", policy: PersistencePolicyDisabled, disabled: true},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			choice, err := PersistenceChoiceForPolicy(tc.policy, projectRoot, testRuntimePersistenceStoreFactory)
+			if err != nil {
+				t.Fatalf("PersistenceChoiceForPolicy: %v", err)
+			}
+			store, err := choice.resolve()
+			if err != nil {
+				t.Fatalf("resolve policy choice: %v", err)
+			}
+			if (store == nil) != tc.disabled {
+				t.Fatalf("store nil = %t, want disabled = %t", store == nil, tc.disabled)
+			}
+		})
+	}
+	if _, err := PersistenceChoiceForPolicy(PersistencePolicy("invalid"), projectRoot, testRuntimePersistenceStoreFactory); err == nil {
+		t.Fatal("PersistenceChoiceForPolicy(invalid) error = nil, want validation error")
+	} else if validation, ok := err.(*ValidationError); !ok || validation.Field != "persistence.policy" {
+		t.Fatalf("invalid policy error = %#v, want persistence.policy ValidationError", err)
+	}
+}
+
+func TestPersistenceChoiceForPolicy_DefaultDoesNotCreateProjectDurableSessions(t *testing.T) {
+	t.Parallel()
+	projectRoot := t.TempDir()
+	storeCalls := 0
+	choice, err := PersistenceChoiceForPolicy(
+		"",
+		projectRoot,
+		func(string) (runtimepersist.Store, error) {
+			storeCalls++
+			return testRuntimePersistenceStoreFactory(projectRoot)
+		},
+	)
+	if err != nil {
+		t.Fatalf("PersistenceChoiceForPolicy(default): %v", err)
+	}
+	store, err := choice.resolve()
+	if err != nil {
+		t.Fatalf("resolve default policy: %v", err)
+	}
+	if store != nil {
+		t.Fatal("default persistence policy unexpectedly configured a store")
+	}
+	if storeCalls != 0 {
+		t.Fatalf("default policy store factory calls = %d, want 0", storeCalls)
+	}
+	if _, err := os.Stat(runtimepersist.DirForProjectRoot(projectRoot)); !os.IsNotExist(err) {
+		t.Fatalf("durable persistence path stat error = %v, want not exist", err)
+	}
+}
+
+func TestPersistenceChoiceForPolicy_EnabledCreatesProjectDurableSessions(t *testing.T) {
+	t.Parallel()
+	projectRoot := t.TempDir()
+	choice, err := PersistenceChoiceForPolicy(
+		PersistencePolicyEnabled,
+		projectRoot,
+		testRuntimePersistenceStoreFactory,
+	)
+	if err != nil {
+		t.Fatalf("PersistenceChoiceForPolicy(enabled): %v", err)
+	}
+	store, err := choice.resolve()
+	if err != nil {
+		t.Fatalf("resolve enabled policy: %v", err)
+	}
+	if store == nil {
+		t.Fatal("enabled persistence policy returned a nil store")
+	}
+	sessionID := "dur-sess-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+	if err := store.Save(sessionID, []byte(`{"sessionId":"dur-sess-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"}`)); err != nil {
+		t.Fatalf("Save: %v", err)
+	}
+	if _, err := os.Stat(filepath.Join(runtimepersist.DirForProjectRoot(projectRoot), sessionID+".json")); err != nil {
+		t.Fatalf("enabled persistence snapshot stat error = %v, want exist", err)
+	}
+}
+
+func TestPrepareStartAndPersistenceHelpers(t *testing.T) {
+	t.Parallel()
+	projectRoot := writeSimpleFinalWorkflowProject(t)
+	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
+		ProjectRoot: projectRoot,
+		Persistence: mustTestRuntimePersistenceStore(t, runtimepersist.DirForProjectRoot(projectRoot)),
+	})
+
+	prepared, err := service.prepareStart(StartRequest{
+		RequestID: "req-prepare-start-001",
+		Source: Source{
+			Kind:         factory.WorkflowSourceKindWorkflowName,
+			WorkflowName: "simple-final",
+		},
+		Args: map[string]any{
+			"subject": "workflows",
+			"count":   1,
+			"prefix":  "you",
+		},
+	})
+	if err != nil {
+		t.Fatalf("prepareStart: %v", err)
+	}
+	if prepared.SourceRef == "" || prepared.SourceContent == "" || prepared.TupleHash == "" {
+		t.Fatalf("prepared start = %#v, want resolved source fields", prepared)
+	}
+
+	terminal, err := service.executeImmediateSyncSession(context.Background(), prepared.Request, prepared.ResolvedSource, prepared.SourceContent, policyResolutionFromPrepared(prepared), "dur-sess-prepare-001")
+	if err != nil {
+		t.Fatalf("executeImmediateSyncSession: %v", err)
+	}
+	sessionID, err := NewDurableSessionID(testSessionIDGenerator)
+	if err != nil {
+		t.Fatalf("NewDurableSessionID: %v", err)
+	}
+	terminal.session.SessionID = sessionID
+	terminal.result.SessionID = sessionID
+	if terminal.session.Status != LifecycleStatusSucceeded {
+		t.Fatalf("terminal session status = %q, want SUCCEEDED", terminal.session.Status)
+	}
+
+	if err := service.persistTerminalSessionState(terminal); err != nil {
+		t.Fatalf("persistTerminalSessionState: %v", err)
+	}
+	loaded, err := service.snapshotSessionState(sessionID)
+	if err != nil {
+		t.Fatalf("snapshotSessionState(load persisted): %v", err)
+	}
+	if loaded.session.SessionID != sessionID {
+		t.Fatalf("loaded sessionID = %q, want %q", loaded.session.SessionID, sessionID)
+	}
+	if loaded.result.ResultStatus != ResultStatusFinal {
+		t.Fatalf("loaded result status = %q, want FINAL", loaded.result.ResultStatus)
+	}
+}
+
+func TestJavaScriptRuntimeService_ProjectRootAloneDoesNotEnablePersistence(t *testing.T) {
+	t.Parallel()
+	projectRoot := writeSimpleFinalWorkflowProject(t)
+	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{ProjectRoot: projectRoot})
+
+	if _, err := service.StartSync(context.Background(), StartRequest{
+		RequestID: "req-runtime-no-implicit-persistence-001",
+		Source: Source{
+			Kind:         factory.WorkflowSourceKindWorkflowName,
+			WorkflowName: "simple-final",
+		},
+	}); err != nil {
+		t.Fatalf("StartSync: %v", err)
+	}
+
+	if _, err := os.Stat(runtimepersist.DirForProjectRoot(projectRoot)); !os.IsNotExist(err) {
+		t.Fatalf("durable persistence path stat error = %v, want not exist", err)
+	}
+}
+
+func TestPersistAndMetadataNoOpBranches(t *testing.T) {
+	t.Parallel()
+	if err := (&JavaScriptRuntimeService{}).persistTerminalSessionState(runtimeSessionState{}); err != nil {
+		t.Fatalf("persistTerminalSessionState(no dir) = %v, want nil", err)
+	}
+
+	projectRoot := t.TempDir()
+	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
+		ProjectRoot: projectRoot,
+		Persistence: mustTestRuntimePersistenceStore(t, runtimepersist.DirForProjectRoot(projectRoot)),
+	})
+
+	if err := service.persistTerminalSessionState(runtimeSessionState{
+		session: SessionReadResult{SessionID: "dur-sess-eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee", Status: LifecycleStatusRunning},
+	}); err != nil {
+		t.Fatalf("persistTerminalSessionState(non-terminal) = %v, want nil", err)
+	}
+	if err := service.persistTerminalSessionState(runtimeSessionState{
+		session: SessionReadResult{Status: LifecycleStatusSucceeded},
+	}); err != nil {
+		t.Fatalf("persistTerminalSessionState(empty session id) = %v, want nil", err)
+	}
+
+	metadata := workflowMetadataFromResolved(ResolvedSource{SourceRef: "fallback-ref"}, StartRequest{})
+	if metadata["name"] != "fallback-ref" {
+		t.Fatalf("fallback metadata name = %#v, want fallback-ref", metadata["name"])
+	}
+}
